### 2019-08-14 #1646 @whedon
[PRE REVIEW]: FluxCalR: a R package for calculating CO2 and CH4 fluxes from static chambers
**Submitting author:** @junbinzhao (<a href="http://orcid.org/0000-0001-5142-4901">Junbin Zhao</a>)
**Repository:** <a href="https://github.com/junbinzhao/FluxCalR/" target ="_blank">https://github.com/junbinzhao/FluxCalR/</a>
**Version:** 0.1.0.
