### 2019-08-16 #1646 @junbinzhao
Dear editor,

Since the application of the package will be mainly in the field of ecological observation, it would be nice to have it reviewed by e.g., @ethanwhite or @jhollist in the reviewer list who has an ecology background.

Thank you!

Junbin
