### 2019-08-16 #1646 @danielskatz
@karthik  - any chance you want to edit this submission?  I can't remember what your JOSS-hiatus dates are, but it otherwise would be perfect for you.
