### 2019-08-26 #1646 @usethedata
Sorry to be slow in responding -- I'm two weeks out from moving and there are lots of tasks needing attention.  Yes, I will review and I have some thoughts for some reviewers to recruit.
