### 2019-08-28 2b866c7
Add community guidelines
