### 2019-08-28 #1646 @junbinzhao
> @junbinzhao One thing I see on a quick review is that contribution guidelines are missing. See https://joss.readthedocs.io/en/latest/review_criteria.html for what we're looking for.

Community guidelines are added. Thanks!
