### 2019-09-03 #1646 @jhollist
Wouldn't be able to get to this review until late September.  Could
probably have it completed during the week of Sep 30.  If that works for
all, then I can do the review.  If you need it sooner, and find other
reviewers, then by all means go that route!

On Tue, Aug 27, 2019 at 6:08 PM Bruce Wilson <notifications@github.com>
wrote:

> @jhollist <https://github.com/jhollist> -- Given your R and ecology
> experience, would you be willing to review this Journal of Open Source
> Software submission? Thanks for the reviews you've already done.
>
