### 2019-09-19 #1646 @molgor
Hi guys,

Thank you for your email.
I'm afraid I can't review any paper at the moment. I'm in the last month(s)
of my PhD and I'm still have lots of things to do.
I'll be happy to review papers after January.

All the best and good luck

Juan

On Thu, Sep 19, 2019 at 12:32 PM Arfon Smith <notifications@github.com>
wrote:

> @molgor <https://github.com/molgor> @ha0ye <https://github.com/ha0ye>
> @GISKid <https://github.com/GISKid> @SteveViss
> <https://github.com/SteveViss> - would any of you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
