### 2019-09-19 #1646 @arfon
> I am able, though please check with @kthyng re: #1550. It sounded like maybe @ethanwhite is committed to that submission, in which case, I'd be happy to be redirected over to this one.

👍 got it, thanks. @kthyng - are you OK for this submission to move forward with @ha0ye as a reviewer?
