### 2019-09-19 #1646 @kthyng
Yes, I still need to hear from @ethanwhite but @ha0ye had a bit of a conflict anyway so better not to have him as a reviewer. So, go for it!
