### 2019-10-01 junbinzhao/FluxCalR#2 @jhollist
Hollister JOSS Review comments
First, nice job on the package!

Second,  my review comments are detailed below for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1751>)  My comments below should match closely with the review checklist and in cases where I didn't check an item taking care of the comments here will satisfy that item.

### General Comments

- Swap tidyverse imports with specific packages.  Tidyverse is big and will install many packages not needed by FluxCalR and (perhaps) not wanted by your users. The specific ones I saw on a quick scan of your code were: magrittr, dplyr, lubridate, and assertthat.  Just specify those specifically in Imports:

### Functionality

- When using FluxCal I get following Warning message.

> Warning message:
> In x11(14, 10) :
>   Calls like 'x11(w, h)' are unsafe and should be replaced by 'x11(width=w, height=h)'

Lines 82 and 218 seem to be the source of this message.  Clean up the code to get rid of this warning message.

### Documentation
note: Much of the documentation is included in the required R documentation.  There is some of this also in the README.  I am not sure what the JOSS requirements are for amount of info in the README vs language specific documentation.

- Statement of need:
  - Did not see a clear statement of need in either the README or the R Package documentation.  Why is this package needed?  Are there other standard ways to load and calculate fluxes?  If not, then there's the need but if so, why an R package specifically.  I don't doubt the need for this kind of package, but needs to be stated.

- Installation instructions:
    - switch install directions from devtools::install_github() to remotes::install_github()

- Example usage:

  - In function level doc, but not in README?  See note above on clarification from JOSS

- Functionality documentation:
  - Function level documentation is good enough and the examples worked, but additional details on what is going on would be better.  I think the package documentation would be improved by a vignette that deals with the flux calculations a little bit more, how to select the END times, etc.

- Automated tests:
  - No clear tests were included in the package.  These need to be added.  A good first place to look for guidance on this is <http://r-pkgs.had.co.nz/tests.html>

- Community guidelines:
  - Community guidelines were included in the README.  Also including a CONTRIBUTING.md is nice as some may look for guidelines there.  An example of a good one is at https://github.com/ropensci/drake/blob/master/CONTRIBUTING.md.  Not required but suggested.

### Software Paper

- State of the field: Didn't see much discussion of how these calculations are typically done.  A little bit more about other methods that researchers use or if those aren't standardized point that out as well.

- Quality of writing:  Just a few small typos/editorial things:
  - Paragraph 1, line 2: "chamber is" to "chambers are" or maybe "The static chamber"
  - Paragraph 1, line 3: "approach" to "approaches"
  - Paragraph 1, lines 7-11: This sentence is awkward.  Should probably be broken up into two sentences.
  - Paragraph 1, line 11: "calculate" doesn't seem right here.  Maybe change to "transfrom"?

  - Paragraph 2, line 1: change "with static" to "with the static"
  - Paragraph 2, line 3: change "multiple flexible" to "multiple, flexible"
  - Paragraph 2, line 3: change "of" to "for"
  - Paragraph 2, line 7: superscript on R^2^?
  - Paragraph 2, line 7: change "within defined" to "within the defined"

  - Paragraph 3, line 1: replace "e.g.," with ", for example,"
  - Paragraph 3, line 7-8: replace "the post-data-analysis with the ready-calculated
fluxes" with "data analysis with the ready calculated fluxes"
