### 2019-10-01 #1751 @jhollist
First, nice job on the package.

Added my checklist to the original post in this thread and in the items I did not check, I provided details over at https://github.com/junbinzhao/FluxCalR/issues/2.

Any questions let me know!
