### 2019-10-07 #1751 @ha0ye
Nice work on a useful tool. The post-calculation graph was nicely done.

There were some checklist items that @jhollist already covered, so I only added comments where I had additional thoughts or suggestions for improvements, over in junbinzhao/FluxCalR#3
