### 2019-10-08 #1751 @arfon
> @arfon & @jhollist & @ha0ye , Thank you for looking at the package! I roughly went through the comments and they look very helpful. I will find time to revise it this month. But since my calendar is bit tight recently, I'll try to finish revising early next month (November).

Sounds good. I'll set a reminder for us all.
