+++ DESCRIPTION
@@ -10,7 +10,7 @@ License: GPL-2
 Encoding: UTF-8
 LazyData: true
 Imports: tidyverse, lubridate, assertthat,magrittr
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.1
 Suggests: knitr,
     rmarkdown
 VignetteBuilder: knitr
