+++ paper.bib
@@ -8,7 +8,7 @@ year = "2004",
 issn = "0168-1923",
 doi = "https://doi.org/10.1016/j.agrformet.2003.12.001",
 url = "http://www.sciencedirect.com/science/article/pii/S0168192303002934",
-author = "Jukka Pumpanen and Pasi Kolari and Hannu Ilvesniemi and Kari Minkkinen and Timo Vesala and Sini Niinistö and Annalea Lohila and Tuula Larmola and Micaela Morero and Mari Pihlatie and Ivan Janssens and Jorge Curiel Yuste and José M Grünzweig and Sascha Reth and Jens-Arne Subke and Kathleen Savage and Werner Kutsch and Geir Østreng and Waldemar Ziegler and Peter Anthoni and Anders Lindroth and Pertti Hari",
+author = "Jukka Pumpanen and Pasi Kolari and Hannu Ilvesniemi and Kari Minkkinen and Timo Vesala and Sini NiinistÃ¶ and Annalea Lohila and Tuula Larmola and Micaela Morero and Mari Pihlatie and Ivan Janssens and Jorge Curiel Yuste and JosÃ© M GrÃ¼nzweig and Sascha Reth and Jens-Arne Subke and Kathleen Savage and Werner Kutsch and Geir Ã˜streng and Waldemar Ziegler and Peter Anthoni and Anders Lindroth and Pertti Hari",
 keywords = "Chamber, Diffusion, Porosity, Soil CO efflux, Turbulence"
 }
 
