+++ R/FluxCal.R
@@ -30,7 +30,8 @@
 #' # load the data
 #' Flux <- Load_LGR(example_data)
 #' # manually select time cues and calculate the fluxes over a 3-minute window
-#' FluxCal(data = Flux, t = 3, vol = 208, Area = 0.26)
+#' Flux_Feb_17 <- FluxCal(data = Flux, t = 3, vol = 208, Area = 0.26)
+#' Flux_Feb_17
 #' # input the time cues as a file and calculate the fluxes over a 3-minute window
 #' Time_Ta <- system.file("extdata", "Time & Ta.csv", package = "FluxCalR") # directory of the file with time cues and Ta
 #' Flux_Feb_17 <- FluxCal(data = Flux, t = 3,
@@ -135,7 +136,11 @@ FluxCal <- function(data,
   colnames(R2.CO2) <- c("R2_CO2","Index_CO2","Slope_CO2","p_CO2","Flux","Date","time_CO2")
   for (a in 1:length(In)){
     for (b in In[a]:c(In[a]-0.5*t*60/f)){ # select 0.5 of the window width as testing for largest R2
-      Slm <- summary(lm(flux$X.CO2.d_ppm[(b-t*60/f):b]~flux$Row[(b-t*60/f):b]))
+      Slm <- try(summary(lm(flux$X.CO2.d_ppm[(b-t*60/f):b]~flux$Row[(b-t*60/f):b])),silent = TRUE)
+      if (class(Slm)=="try-error"){ # if no CO2 data is provided
+        R2.CO2[a,1:7] <- NA
+        Ta_CO2[a] <- NA
+      } else {
       if (Slm$r.squared > R2.CO2[a,1]) {
         R2.CO2[a,1] <- Slm$r.squared
         # output the time at the start of the slope
@@ -149,6 +154,7 @@ FluxCal <- function(data,
                               floor(lubridate::second(flux$Time[(b-t*60/f)])))
         Ta_CO2[a] <- mean(flux$AmbT_C[(b-t*60/f):b])
       }
+      }
     }
   }
 
@@ -158,7 +164,11 @@ FluxCal <- function(data,
   colnames(R2.CH4) <- c("R2_CH4","Index_CH4","Slope_CH4","p_CH4","Flux","Date","time_CH4")
   for (a in 1:length(In)){
     for (b in In[a]:c(In[a]-0.5*t*60/f)){
-      Slm <- summary(lm(flux$X.CH4.d_ppm[(b-t*60/f):b]~flux$Row[(b-t*60/f):b]))
+      Slm <- try(summary(lm(flux$X.CH4.d_ppm[(b-t*60/f):b]~flux$Row[(b-t*60/f):b])),silent=TRUE)
+      if (class(Slm)=="try-error"){ # if no CH4 data is provided
+        R2.CH4[a,1:7] <- NA
+        Ta_CH4[a] <- NA
+      } else {
       if (Slm$r.squared > R2.CH4[a,1]) {
         R2.CH4[a,1] <- Slm$r.squared
         R2.CH4[a,2] <- b
@@ -170,6 +180,7 @@ FluxCal <- function(data,
                               lubridate::minute(flux$Time[(b-t*60/f)]),":",
                               floor(lubridate::second(flux$Time[(b-t*60/f)])))
         Ta_CH4[a] <- mean(flux$AmbT_C[(b-t*60/f):b])
+      }
       }
     }
   }
[…]
