+++ DESCRIPTION
@@ -10,8 +10,8 @@ Depends: R (>= 3.4.1)
 License: MIT
 Encoding: UTF-8
 LazyData: true
-Imports: tidyverse, lubridate, assertthat,magrittr
-RoxygenNote: 6.0.1
+Imports: tidyverse, lubridate, assertthat, magrittr
+RoxygenNote: 6.1.1
 Suggests: knitr,
     rmarkdown
 VignetteBuilder: knitr
