+++ R/FluxCal.R
@@ -10,18 +10,18 @@
 #'
 #' @param data A dataframe generated by the function "Load_LGR" or "Load_other".
 #' @param t A number indicates the window width for the flux calculation, unit: minutes.
-#' @param spt A number indicates the number of graphs to split the CO2 time series into for MANUALLY selecting the time cues. Cutting CO2 time series into several sections will aid the visualization and selection of the time cues when the data is too long. Default: 1. This argument is used only when no data is provided for the "Time_keys" argument.
+#' @param vol A number indicates volume of the chamber; unit: dm^3 or L.
+#' @param Area A number indicates base area of the chamber; unit: m^2.
+#' @param spt An integer indicates the number of graphs to split the CO2 time series into for MANUALLY selecting the time cues. Cutting CO2 time series into several sections will aid the visualization and selection of the time cues when the data is too long. Default: 1. This argument is used only when no data is provided for the "Time_keys" argument.
 #' @param Time_keys A string inidicates the directory and the name of the csv file that includes start or end time (HH:MM:SS) of each flux measurement.The header for the time must be either "Start" or "End". (see example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
 #' Default: if no file is provided (NULL), a graph window will pop up to allow manually selecting the END time, which could be normally recognized as the local "peaks" or "valleys" of the time series.
 #' @param Type_keys Either "start" or "end" (string) to indicate if start or end time will be used as the cues; default "start". This will only valid if a file name is provided for argument "Time_keys".
-#' @param vol A number indicates volume of the chamber; unit: dm^3 or L.
 #' @param Ta A name (string) of a file contains a column "Ta" with the air temperature values (ideally, this is temperature measured inside of the chamber during the flux measurement; unit: degree C. See example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata). The length of Ta within the file must be the same as the number of flux measurements.
 #' Default: if no file name is provided (NULL), the average ambient air temperature measured by the analyzer or Ta input from function "Load_other" will be used.
-#' @param Area A number indicates base area of the chamber; unit: m^2.
 #' @param output_d A string includes output directory and file name (.csv) to export the calculated fluxes. Default: a file named "Flux_output.csv" with calculated fluxes will be created under the current work directory. FALSE, do not create the file.
 #' @param ylim_CO2 A numeric vector of length 2, giving the y-axis scale range for CO2 concentration (ppm) for plotting purpose. If not specified (default), it will be set based on the CO2 range of the entire dataset.
 #' @param ylim_CH4 A numeric vector of length 2, giving the y-axis scale range for CH4 concentration (ppm) for plotting purpose. If not specified (default), it will be set based on the CH4 range of the entire dataset.
-#'
+#' @param digits An integer indicates the number of decimal digits to be used for the calculated flux and R2
[…]
