### 2019-10-11 05bc213
created a function dedicate to extract the time cues
