### 2019-10-11 34972bb
the time cue function is working! Need documentations
