+++ R/Select time cues.R
@@ -6,51 +6,48 @@
 #' @export
 ## function to manually select the time cues
 Sel_cue <- function(data,
-                    Flux = "X.CO2.d_ppm",
+                    Flux = "CO2",
                     cue = "End",
                     spt = 1,
                     ylim = NULL,
                     save = "Time_cue.csv"){
   # define the pipe from the package "magrittr"
   `%>%` <- magrittr::`%>%`
-
   # creating a index for spiliting the window
   nr <- nrow(data) # number of rows
-  # add one column as row index and one as time (HH:MM:SS)
+  # add one column as row index, one as time (HH:MM:SS) and one as the flux to be ploted (either "CO2" or "CH4")
   data <- cbind(Row=row(data)[,1],data) %>%
-    dplyr::mutate(time=paste(lubridate::hour(Time),sprintf("%02d",lubridate::minute(Time)),sprintf("%02d",floor(lubridate::second(Time))),sep = ":"))
+    dplyr::rowwise() %>%
+    dplyr::mutate(flux_plot=ifelse(Flux=="CO2",X.CO2.d_ppm,X.CH4.d_ppm))
 
   ######## select the points to separate the measurements
-  # plot CO2 vs index for locating the peaks and valleys
+  # define the ylab and ylim
+  ylab <- paste0(ifelse(Flux=="CO2","CO2","CH4"),
+                 " in ppm")
+  if (is.null(ylim)){
+    ylim <- range(data$flux_plot)
+  } else {
+    ylim <- ylim
+  }
+
+  # plot flux vs time for locating the peaks and valleys
   In <- c() # create a variable as Index
   for (i in 1:spt){
     x11(width = 16,height = 10)
-    if (is.null(ylim_CO2)){
-      with(data[c(((i-1)*nr/spt)+1):c(i*nr/spt),],
-           plot(Time,X.CO2.d_ppm,
-                ylab = "CO2 in ppm",
-                main = "Click on the peaks and valleys",
-                cex.main=2.5, col.main="red",
-                cex = 0.8, xaxt="n"
-           ))
-    } else {
-      with(data[c(((i-1)*nr/spt)+1):c(i*nr/spt),],
-           plot(Time,X.CO2.d_ppm,
-                ylab = "CO2 in ppm",
-                main = "Click on the peaks and valleys",
-                cex.main=2.5, col.main="red",
-                ylim = ylim_CO2, cex = 0.8, xaxt="n"
-           ))
-    }
-
-
+    with(data[c(((i-1)*nr/spt)+1):c(i*nr/spt),],
+         plot(Time,flux_plot,
+              ylab = ylab,
+              main = "Click on the peaks and valleys",
+              cex.main=2.5, col.main="red",
+              ylim = ylim, cex = 0.8, xaxt="n"
+         ))
     # add time interval ticks
     a <- lubridate::pretty_dates(data$Time[c(((i-1)*nr/spt)+1):c(i*nr/spt)],10)
     # which(minute(data$Time) %in% c(0,15,30,45) & floor(second(data$Time)) %in% c(0,1))
     axis.POSIXct(1, at= a,format = "%H:%M")
     abline(v=a, lty="dotted",col="grey")
-    In_t <- with(data[c(((i-1)*nr/spt)+1):c(i*nr/spt),],identify(Time,X.CO2.d_ppm))
-    In_t <- In_t + ((i-1)*nr/spt)
+    In_t <- with(data[c(((i-1)*nr/spt)+1):c(i*nr/spt),],identify(Time,flux_plot))
+    In_t <- In_t + ((i-1)*nr/spt) # get the actual location in the whole dataset
     In <- c(In,In_t)
     dev.off()
   }
@@ -63,7 +60,7 @@ Sel_cue <- function(data,
                          End)
[…]
