### 2019-10-11 b2f8b05
added the instruction of clicking
