### 2019-10-14 25cfdca
change the output as the HHMMSS
