+++ man/Sel_cue.Rd
@@ -1,0 +1,53 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/Select time cues.R
+\name{Sel_cue}
+\alias{Sel_cue}
+\title{Select the time cues}
+\usage{
+Sel_cue(data, Flux = "CO2", cue = "End", spt = 1, ylim = NULL,
+  save = "Time_cue.csv")
+}
+\arguments{
+\item{data}{A dataframe generated by the function "Load_LGR" or "Load_other".}
+
+\item{Flux}{A string, either "CO2" (default) or "CH4", indicates the either CO2 or CH4 concentration
+should be plotted for selecting the time cues.}
+
+\item{cue}{A string, either "End" (default) or "Start", indicates whether the end or the start time of each measurement is
+to be selected. "End" is recommended is here, since the end of each measurement is more identifiable which is
+usually associated with a sudden drop/rise in the concentrations due to the removal of chamber.}
+
+\item{spt}{An integer indicates the number of sections to split the CO2/CH4 time series into when plotting (default: 1).
+Cutting CO2/CH4 time series into several sections will aid the visualization and selection of the time cues when the data is too long.}
+
+\item{ylim}{A numeric vector of length 2, giving the y-axis scale range for CO2/CH4 concentrations (ppm).
+If not specified (default), it will be set based on the CO2/CH4 range of the entire dataset.}
+
+\item{save}{A string includes output directory and file name (.csv) to export the extracted time cue data frame.
+Default: a file named "Time_cue.csv" with the extracted time cues will be created under the current work directory.
+The file created can be used in the flux calculation using the function `FluxCal` to get reproducible results.
+FALSE, do not create the file.}
+}
+\value{
+A data frame that includes two columns:
+"No", the number of each measurement;
+"Start" or "End", the time selected as the start or end of each measurement (HH:MM:SS)
+}
+\description{
+This is a function to manually select and extract the start or the end of each measurement as time cues for
+flux calculations. The data frame returned from this function can be assigned to the argument `Time_keys` in the `FluxCal` function.
+Note that the time cue data frame can also be prepared by the user independent of this function following the format of
+"Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
+}
+\examples{
+#### data from LGR
+# get the directory of the example LGR raw data
+example_data1 <- system.file("extdata", "Flux_example_1_LGR.txt", package = "FluxCalR")
+example_data1
+# load the data
+Flux_lgr <- Load_LGR(example_data1)
+# select cues
+Cues <- Sel_cue(Flux_lgr, save = FALSE)
+Cues
+
+}
