+++ NAMESPACE
@@ -3,3 +3,4 @@
 export(FluxCal)
 export(Load_LGR)
 export(Load_other)
+export(Sel_cue)
