### 2019-10-14 1e850a2
range choose to calculate the flux
