+++ R/FluxCal.R
@@ -9,7 +9,7 @@
 #' Based on the time cues and window width provided for the calculation, the function will automatically scan over data that cover 1.5x length of the window width and calculate the fluxes based on the best linear regression (i.e., largest R2). After the calculations are done, a graph with regression lines plotted on the CO2 and/or CH4 concentration time series will pop up for checkup purposes.
 #'
 #' @param data A dataframe generated by the function "Load_LGR" or "Load_other".
-#' @param t A number indicates the window width for the flux calculation, unit: minutes.
+#' @param win A number indicates the window width for the flux calculation, unit: minutes.
 #' @param vol A number indicates volume of the chamber; unit: dm^3 or L.
 #' @param Area A number indicates base area of the chamber; unit: m^2.
 #' @param spt An integer indicates the number of graphs to split the CO2 time series into for MANUALLY selecting the time cues. Cutting CO2 time series into several sections will aid the visualization and selection of the time cues when the data is too long. Default: 1. This argument is used only when no data is provided for the "Time_keys" argument.
@@ -69,88 +69,56 @@
 #' @export
 ## Flux calculation function ---------------
 FluxCal <- function(data,
-                    t,
+                    Cal = "CO2_CH4",
+                    win,
                     vol,
                     Area,
-                    spt=1,
-                    Time_keys = NULL,
-                    Type_keys = "start",
+                    Cue,
+                    Cue_type = "End",
                     Ta = NULL,
+                    Other,
+                    ext = 1.5,
                     output_d = "Flux_output.csv",
-                    ylim_CO2 = NULL,
-                    ylim_CH4 = NULL,
                     digits = 3
 ) {
   # define the pipe from the package "magrittr"
   `%>%` <- magrittr::`%>%`
   # constants
   R_index = 0.08205783 # universal gas constant; unit: L*atm*K^-1*mol^-1
-
-  flux <- data
-
-  # creating a index for spiliting the window
-  nr <- nrow(flux) # number of rows
-
-  f <- round(as.numeric(flux$Time[2]-flux$Time[1])) # calculate the sampling frequency based on the timestamps.
+  # calculate the sampling frequency based on the timestamps (unit: seconds)
+  f <- round(as.numeric(difftime(data$Time[2],data$Time[1],units = "secs")))
 
   # add one column as row index and one as time (HH:MM:SS)
-  flux <- cbind(Row=row(flux)[,1],flux) %>%
-    dplyr::mutate(time=paste(lubridate::hour(Time),sprintf("%02d",lubridate::minute(Time)),sprintf("%02d",floor(lubridate::second(Time))),sep = ":"))
+  data <- cbind(Row=row(data)[,1],data) %>%
+    dplyr::mutate(time=paste(lubridate::hour(Time),
+                             sprintf("%02d",lubridate::minute(Time)),
[…]
