### 2019-10-14 a5592df
finished the flux calculation function within the main function
