+++ R/FluxCal.R
@@ -17,7 +17,7 @@
 #' Default: if no file is provided (NULL), a graph window will pop up to allow manually selecting the END time, which could be normally recognized as the local "peaks" or "valleys" of the time series.
 #' @param Type_keys Either "start" or "end" (string) to indicate if start or end time will be used as the cues; default "start". This will only valid if a file name is provided for argument "Time_keys".
 #' @param Ta A name (string) of a file contains a column "Ta" with the air temperature values (ideally, this is temperature measured inside of the chamber during the flux measurement; unit: degree C. See example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata). The length of Ta within the file must be the same as the number of flux measurements.
-#' Default: if no file name is provided (NULL), the average ambient air temperature measured by the analyzer or Ta input from function "Load_other" will be used.
+#' Default: if no file name is provided (NULL), the average ambient air temperature measured by the LGR analyzer or, if the data measured with other analyzers, Ta input from function "Load_other" will be used.
 #' @param output_d A string includes output directory and file name (.csv) to export the calculated fluxes. Default: a file named "Flux_output.csv" with calculated fluxes will be created under the current work directory. FALSE, do not create the file.
 #' @param ylim_CO2 A numeric vector of length 2, giving the y-axis scale range for CO2 concentration (ppm) for plotting purpose. If not specified (default), it will be set based on the CO2 range of the entire dataset.
 #' @param ylim_CH4 A numeric vector of length 2, giving the y-axis scale range for CH4 concentration (ppm) for plotting purpose. If not specified (default), it will be set based on the CH4 range of the entire dataset.
@@ -79,7 +79,8 @@ FluxCal <- function(data,
                     Other,
                     ext = 1.5,
                     output_d = "Flux_output.csv",
-                    digits = 3
+                    digits = 3,
+                    check_plot = TRUE
 ) {
   # define the pipe from the package "magrittr"
   `%>%` <- magrittr::`%>%`
@@ -94,7 +95,7 @@ FluxCal <- function(data,
                              sprintf("%02d",lubridate::minute(Time)),
                              sprintf("%02d",floor(lubridate::second(Time))),sep = ":"))
 
-  ### get the row index for the moving window in calculation ------
+  ### get the row index for the moving window in calculation ---------
   mov <- ext-1 # the extend of moving
   if (Cue_type == "End"){ ### if end time is given
     In1 <- match(Cue$End,data$time) # start moving
@@ -121,83 +122,63 @@ FluxCal <- function(data,
                 " in the 'Cue' 1 second forward..."))
   }
 
-  ### calculate the max R2 of the regression before local max and min
-  # R2 of CO2 fluxes
-  Ta_CO2 <- vector() # for calculating average ambient temperature
-  R2.CO2 <- data.frame(matrix(0,length(In),7))
-  colnames(R2.CO2) <- c("R2_CO2","Index_CO2","Slope_CO2","p_CO2","Flux","Date","time_CO2")
[…]
