### 2019-10-15 7e39d79
updated the graph options
