+++ R/FluxCal.R
@@ -128,7 +128,7 @@ FluxCal <- function(data,
   # function to calculate R2, slopes and fluxes ----------
   CalFUN <- function(flux = "CO2") {
     dft <- data.frame(matrix(0,nrow(Cue),9)) # for record the data
-    names(dft) <- c("R2","Index","Slope","p","Flux","Date","Start","End","Ta")
+    names(dft) <- c("Num","Date","Start","End","Gas","Slope","R2","Ta","Index")
     ########## 1. calculate the max R2 and slopes of the regression winthin the window with an extended range
     for (a in 1:nrow(Cue)){
       # the window is moving backwards from the end point
@@ -140,25 +140,25 @@ FluxCal <- function(data,
         }
         # if no data is provided
         if (class(Slm)=="try-error"){
-          dft[a,1:8] <- NA
-          dft[a,9] <- NA
+          stop(paste0("Make sure ",flux," data are provided and specified in the correct way (See the help)."))
         } else {
-        if (Slm$r.squared > dft[a,1]) {
-          dft[a,1] <- Slm$r.squared
-          dft[a,2] <- b # output the row index at the start of the slope
-          dft[a,3] <- Slm$coefficients[2]/f
-          dft[a,4] <- Slm$coefficients[8]
-          dft[a,5] <- flux
-          dft[a,6] <- paste0(lubridate::year(data$Time[1]),"-", # the date
+          dft[a,1] <- a # the Number of measurements
+        if (Slm$r.squared > dft[a,"R2"]) {
+          dft[a,2] <- paste0(lubridate::year(data$Time[1]),"-", # the date
                              lubridate::month(data$Time[1]),"-",
                              lubridate::day(data$Time[1]))
-          dft[a,7] <- paste0(lubridate::hour(data$Time[(b-t*60/f)]),":", # the start time of the slope
+          dft[a,3] <- paste0(lubridate::hour(data$Time[(b-t*60/f)]),":", # the start time of the slope
                              lubridate::minute(data$Time[(b-t*60/f)]),":",
                              floor(lubridate::second(data$Time[(b-t*60/f)])))
-          dft[a,8] <- paste0(lubridate::hour(data$Time[b]),":", # the end time of the slope
+          dft[a,4] <- paste0(lubridate::hour(data$Time[b]),":", # the end time of the slope
                              lubridate::minute(data$Time[b]),":",
                              floor(lubridate::second(data$Time[b])))
-          dft[a,9] <- mean(data$AmbT_C[(b-t*60/f):b])
+          dft[a,5] <- flux
+          dft[a,6] <- try(Slm$coefficients[2]/f,silent = TRUE) # slope as against 1s
+          dft[a,7] <- try(Slm$r.squared,silent = TRUE)
+          dft[a,8] <- round(mean(data$AmbT_C[(b-t*60/f):b]),digits=2)
+          dft[a,9] <- b # output the row index at the start of the slope for plotting the graphs
+          # dft[a,4] <- Slm$coefficients[8] # p value
         }
         }
       } # end b loop
@@ -168,16 +168,13 @@ FluxCal <- function(data,
[…]
