### 2019-10-17 c2f8a9b
changed output time format and added Other to the output.
