### 2019-10-17 faeedba
improved the output and plotting for both CO2 and CH4
