### 2019-10-18 28aa84a
added error to each error message
