+++ R/FluxCal.R
@@ -11,7 +11,7 @@
 #' @param data A dataframe generated by the function "Load_LGR" or "Load_other".
 #' @param win A number indicates the window width for the flux calculation, unit: minutes.
 #' @param vol A number indicates volume of the chamber; unit: dm^3 or L.
-#' @param Area A number indicates base area of the chamber; unit: m^2.
+#' @param area A number indicates base area of the chamber; unit: m^2.
 #' @param spt An integer indicates the number of graphs to split the CO2 time series into for MANUALLY selecting the time cues. Cutting CO2 time series into several sections will aid the visualization and selection of the time cues when the data is too long. Default: 1. This argument is used only when no data is provided for the "Time_keys" argument.
 #' @param Time_keys A string inidicates the directory and the name of the csv file that includes start or end time (HH:MM:SS) of each flux measurement.The header for the time must be either "Start" or "End". (see example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
 #' Default: if no file is provided (NULL), a graph window will pop up to allow manually selecting the END time, which could be normally recognized as the local "peaks" or "valleys" of the time series.
@@ -69,14 +69,14 @@
 #' @export
 ## Flux calculation function ---------------
 FluxCal <- function(data,
-                    Cal = "CO2_CH4",
+                    cal = "CO2_CH4",
                     win,
                     vol,
-                    Area,
-                    Cue,
-                    Cue_type = "End",
+                    area,
+                    df_cue,
+                    cue_type = "End",
                     df_Ta = NULL,
-                    Other = NULL,
+                    other = NULL,
                     ext = 1.5,
                     output = "Flux_output.csv",
                     digits = 3,
@@ -85,12 +85,16 @@ FluxCal <- function(data,
                     ylim_CH4 = NULL
 
 ) {
+  # check arguments
+  cal <- match.arg(cal,c("CO2_CH4","CO2","CH4"))
+  cue_type <- match.arg(cue_type,c("End","Start","Start_End"))
   # define the pipe from the package "magrittr"
   `%>%` <- magrittr::`%>%`
-  # constants
-  R_index = 0.08205783 # universal gas constant; unit: L*atm*K^-1*mol^-1
   # calculate the sampling frequency based on the timestamps (unit: seconds)
   f <- round(as.numeric(difftime(data$Time[2],data$Time[1],units = "secs")))
+  # constants
+  R_index <- 0.08205783 # universal gas constant; unit: L*atm*K^-1*mol^-1
+  win_f <- win*60/f # the number of rows for the defined window
 
   # add one column as row index and one as time (HH:MM:SS)
   data <- cbind(Row=row(data)[,1],data) %>%
@@ -100,18 +104,18 @@ FluxCal <- function(data,
 
   ### get the row index for the moving window in calculation ---------
   mov <- ext-1 # the extend of moving
-  if (Cue_type == "End"){ ### if end time is given
-    In1 <- match(Cue$End,data$time) # start moving
-    In2 <- In1-mov*win*60/f # end of moving (backwards)
+  if (cue_type == "End"){ ### if end time is given
+    In1 <- match(df_cue$End,data$time) # start moving
[…]
