+++ R/Load.R
@@ -3,7 +3,7 @@
 #' @description Function used to load raw data from LGR Ultraportable Gas Analyzers to feed the "data" argument in the function "FluxCal()".
 #'
 #' @param file A string contains the directory path and file name of the raw data. The data should be provided as what is exported from the LGR analyzer.
-#' @param Time_format A string indicates the format of timestamps. Default: "mdy_hms". It can be changed into "dmy_hms" or "ymd_hms", depending on the data timestamps from the analyzer.
+#' @param Time_format A string indicates the format of timestamps. Default: "mdy_HMS". It can be changed into "dmy_HMS" or "ymd_HMS", depending on the data timestamps from the analyzer.
 #'
 #' @return A dataframe with all data in the right format for function "FluxCal()".
 #'
@@ -12,25 +12,23 @@
 #' example_data <- system.file("extdata", "Flux_example_1_LGR.txt", package = "FluxCalR")
 #' example_data
 #' # load the data
-#' Flux <- Load_LGR(example_data)
+#' Flux <- LoadLGR(example_data)
 #'
 #' @export
 ## Function for loading data from LGR ----------
-Load_LGR <- function(file,
-                     Time_format="mdy_hms"
+LoadLGR <- function(file,
+                    time_format="mdy_HMS"
 ){
+  # argument check
+  time_format <- match.arg(time_format,c("mdy_HMS","dmy_HMS","ymd_HMS"))
   # Load the raw data after deleting the text massages at the beginning and the end of each raw data file generate by LGR
   flux <- read.table(file, sep= ",", header = TRUE,skip = 1,fill = TRUE,stringsAsFactors = FALSE)
   flux <- subset(flux,!is.na(flux[,2])) # remove the comments generated by LGR
   # replace the time with "lubridated" time with different format
-  if (Time_format=="mdy_hms"){
-    flux$Time <- lubridate::mdy_hms(flux$Time)
-  } else {
-    if (Time_format=="dmy_hms"){
-      flux$Time <- lubridate::dmy_hms(flux$Time)
-    } else {
-      flux$Time <- lubridate::ymd_hms(flux$Time) # with Time_format = "ymd_hms"
-    }
+  flux$Time <- try(lubridate::parse_date_time(flux$Time,orders = time_format),silent = TRUE)
+  # error if time column is note correctly referred to
+  if (class(flux$Time)[1]=="try-error"){
+    stop("Error: check if the 'time_format' argument is correct!")
   }
   return(flux)
 }
@@ -41,11 +39,11 @@ Load_LGR <- function(file,
 #' @description Function used to load raw data output from other similar analyzers (e.g. LICOR) to feed the "data" argument in function "FluxCal()".
 #'
 #' @param file A string contains the directory path and file name of the raw data.
-#' @param Time A string indicates column name of the timestamps in the data file.
-#' @param Time_format A string indicates the format of timestamps. Default: "mdy_hms".
-#' It can be changed into "dmy_hms" or "ymd_hms", depending on the data timestamp from the analyzer.
-#' In case date and time are recorded separately, "hms" can be used here and add the date to the "Date" argument.
-#' @param Date_ms A string indicates date of the measurements in the format of "mm/dd/yyyy". This argument is used only when date is not included in the timestamps.
+#' @param time A string indicates column name of the timestamps in the data file.
+#' @param time_format A string indicates the format of timestamps. Default: "mdy_HMS".
+#' It can be changed into "dmy_HMS" or "ymd_HMS", depending on the data timestamp from the analyzer.
[…]
