+++ R/Select time cues.R
@@ -7,7 +7,7 @@
 #'
 #' @param data A dataframe generated by the function "Load_LGR" or "Load_other".
 #'
-#' @param Flux A string, either "CO2" (default) or "CH4", indicates the either CO2 or CH4 concentration
+#' @param flux A string, either "CO2" (default) or "CH4", indicates the either CO2 or CH4 concentration
 #' should be plotted for selecting the time cues.
 #'
 #' @param cue A string, either "End" (default) or "Start", indicates whether the end or the start time of each measurement is
@@ -35,19 +35,22 @@
 #' example_data1 <- system.file("extdata", "Flux_example_1_LGR.txt", package = "FluxCalR")
 #' example_data1
 #' # load the data
-#' Flux_lgr <- Load_LGR(example_data1)
+#' Flux_lgr <- LoadLGR(example_data1)
 #' # select cues
-#' Cues <- Sel_cue(Flux_lgr, save = FALSE)
+#' Cues <- SelCue(Flux_lgr, save = FALSE)
 #' Cues
 #'
 #' @export
 ## function to manually select the time cues --------
-Sel_cue <- function(data,
-                    Flux = "CO2",
-                    cue = "End",
-                    spt = 1,
-                    ylim = NULL,
-                    save = "Time_cue.csv"){
+SelCue <- function(data,
+                   flux = "CO2",
+                   cue = "End",
+                   spt = 1,
+                   ylim = NULL,
+                   save = "Time_cue.csv"){
+  # argument check
+  flux <- match.arg(flux,c("CO2","CH4"))
+  cue <- match.arg(cue,c("End","Start"))
   # define the pipe from the package "magrittr"
   `%>%` <- magrittr::`%>%`
   # creating a index for spiliting the window
@@ -55,11 +58,11 @@ Sel_cue <- function(data,
   # add one column as row index, one as time (HH:MM:SS) and one as the flux to be ploted (either "CO2" or "CH4")
   data <- cbind(Row=row(data)[,1],data) %>%
     dplyr::rowwise() %>%
-    dplyr::mutate(flux_plot=ifelse(Flux=="CO2",X.CO2.d_ppm,X.CH4.d_ppm))
+    dplyr::mutate(flux_plot=ifelse(flux=="CO2",X.CO2.d_ppm,X.CH4.d_ppm))
 
   ######## select the points to separate the measurements
   # define the ylab and ylim
-  ylab <- paste0(ifelse(Flux=="CO2","CO2","CH4"),
+  ylab <- paste0(ifelse(flux=="CO2","CO2","CH4"),
                  " in ppm")
   if (is.null(ylim)){
     ylim <- range(data$flux_plot)
