### 2019-10-21 805f96e
updated the documentation for the loading and time cue selecting functions
