+++ man/LoadLGR.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/Load.R
+\name{LoadLGR}
+\alias{LoadLGR}
+\title{Load LGR data}
+\usage{
+LoadLGR(file, time_format = "mdy_HMS")
+}
+\arguments{
+\item{file}{A string contains the directory path and file name of the raw data.
+The data should be provided as what is exported from the LGR analyzer.}
+
+\item{Time_format}{A string indicates the format of timestamps. Default: "mdy_HMS".
+It also takes the format "dmy_HMS" or "ymd_HMS", depending on the timestamps output from the analyzer.}
+}
+\value{
+A data frame with all data in the right format for the functions `SelCue()` and `FluxCal()`.
+}
+\description{
+A function used to load raw data from LGR Ultraportable Gas Analyzers to be called in the functions
+`SelCue()` and `FluxCal()`. This function removes the extra lines without data at the end of the data file output
+by LGR analyzer and convert the timestamps into POSIXct for further processing.
+}
+\examples{
+# get the directory of the example LGR raw data
+example_data <- system.file("extdata", "Flux_example_1_LGR.txt", package = "FluxCalR")
+# load the data
+Flux <- LoadLGR(example_data)
+
+}
