+++ NAMESPACE
@@ -1,6 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
 export(FluxCal)
-export(Load_LGR)
-export(Load_other)
-export(Sel_cue)
+export(LoadLGR)
+export(LoadOther)
+export(SelCue)

[…]
