+++ R/FluxCal.R
@@ -1,27 +1,47 @@
 #' @title Calculate CO2 and CH4 gas fluxes
 #'
-#' @description Calculate CO2 and CH4 gas fluxes from the data loaded by the function "Load_LGR()" or "Load_other()". The function "FluxCal()" calculates CO2 and/or CH4 flux rates based on the time cues provided for each measurement (i.e., either start or end time). Two options are available to input the time cues:
+#' @description A function to calculate CO2 and CH4 gas fluxes from the data loaded by the function `LoadLGR()` or `Load_other()`.
+#' It takes a time cue data frame (argument `df_cue`), either created by the function `SelCue()` or prepared by the user following the format of
+#' "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata, to separate the
+#' measurements and calculate the fluxes for all measurements at once.
+#' Note that the header for the time cue column must be either *"Start"* or *"End"*.
+#' Based on the time cues and window width provided for the calculation, the function will automatically scan over data that cover
+#'  1.5x (default) length of the window and calculate the fluxes based on the best linear regression (i.e., largest R2).
+#'  After the calculations are done, a graph with regression lines plotted on the CO2 and/or CH4 concentration time series can be
+#'  drawn for checkup purposes.
 #'
-#' (1) (default) after executing the function, manually clicking on a pop-up graph with CO2 concentration time series to choose the END time, which could be identified as the local "peaks" or "valleys" in the time series; or
-#'
-#' (2) loading a file (.csv) into the argument "Time_keys" with times (HH:MM:SS) indicating start or end of each flux measurement. The header for the time must be either "Start" or "End". (see example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
-#'
-#' Based on the time cues and window width provided for the calculation, the function will automatically scan over data that cover 1.5x length of the window width and calculate the fluxes based on the best linear regression (i.e., largest R2). After the calculations are done, a graph with regression lines plotted on the CO2 and/or CH4 concentration time series will pop up for checkup purposes.
-#'
-#' @param data A dataframe generated by the function "Load_LGR" or "Load_other".
+#' @param data A dataframe generated by the function `LoadLGR` or `LoadOther`.
+#' @param cal A string, either "CO2_CH4" (default),"CO2" or "CH4", indicates which gas flux it is calculated for.
 #' @param win A number indicates the window width for the flux calculation, unit: minutes.
 #' @param vol A number indicates volume of the chamber; unit: dm^3 or L.
 #' @param area A number indicates base area of the chamber; unit: m^2.
-#' @param spt An integer indicates the number of graphs to split the CO2 time series into for MANUALLY selecting the time cues. Cutting CO2 time series into several sections will aid the visualization and selection of the time cues when the data is too long. Default: 1. This argument is used only when no data is provided for the "Time_keys" argument.
-#' @param Time_keys A string inidicates the directory and the name of the csv file that includes start or end time (HH:MM:SS) of each flux measurement.The header for the time must be either "Start" or "End". (see example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
-#' Default: if no file is provided (NULL), a graph window will pop up to allow manually selecting the END time, which could be normally recognized as the local "peaks" or "valleys" of the time series.
[…]
