+++ man/SelCue.Rd
@@ -36,8 +36,8 @@ A data frame that includes two columns:
 }
 \description{
 This is a function to manually select and extract the start or the end time of each measurement as time cues by
-clicking the corresponding measurement points on the interactive graphs generated by the function. The data frame returned
-from this function can be assigned to the argument `df_cue` in the `FluxCal()` function.
+clicking the corresponding start or end measurement points on the interactive graphs generated by the function.
+The data frame returned from this function can be assigned to the argument `df_cue` in the `FluxCal()` function.
 Note that the time cue data frame can also be prepared by the user independent of this function following the format of
 "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
 }
