+++ man/FluxCal.Rd
@@ -5,12 +5,14 @@
 \title{Calculate CO2 and CH4 gas fluxes}
 \usage{
 FluxCal(data, cal = "CO2_CH4", win, vol, area, df_cue,
-  cue_type = "End", df_Ta = NULL, other = NULL, ext = 1.5,
+  cue_type = "End", other = NULL, df_Ta = NULL, ext = 1.5,
   output = "Flux_output.csv", digits = 3, check_plot = TRUE,
   ylim_CO2 = NULL, ylim_CH4 = NULL)
 }
 \arguments{
-\item{data}{A dataframe generated by the function "Load_LGR" or "Load_other".}
+\item{data}{A dataframe generated by the function `LoadLGR` or `LoadOther`.}
+
+\item{cal}{A string, either "CO2_CH4" (default),"CO2" or "CH4", indicates which gas flux it is calculated for.}
 
 \item{win}{A number indicates the window width for the flux calculation, unit: minutes.}
 
@@ -18,35 +20,55 @@ FluxCal(data, cal = "CO2_CH4", win, vol, area, df_cue,
 
 \item{area}{A number indicates base area of the chamber; unit: m^2.}
 
-\item{digits}{An integer indicates the number of decimal digits to be used for the calculated flux and R2}
+\item{df_cue}{A data frame that includes "Start" and/or "End" time (HH:MM:SS) of each measurement.
+The header for the time must be *"Start"* or *"End"*. This data frame can either be created by the function `SelCue()`
+or be prepared by the user
+(see example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).}
 
-\item{ylim_CO2}{A numeric vector of length 2, giving the y-axis scale range for CO2 concentration (ppm) for plotting purpose. If not specified (default), it will be set based on the CO2 range of the entire dataset.}
+\item{cue_type}{A string, either "Start", "End" (default) or "Start_End", indicates if start, end or both time in the
+data frame assigned to `df_cue` will be used as the cues . When "Start_End" is chosen, both "Start" and "End" columns have to
+be present in the data frame for `df_cue` and the flux with the largest R2 within the range will be calculated without
+considering the argument `ext`.}
 
-\item{ylim_CH4}{A numeric vector of length 2, giving the y-axis scale range for CH4 concentration (ppm) for plotting purpose. If not specified (default), it will be set based on the CH4 range of the entire dataset.}
+\item{other}{A vector of strings indicates the names of other columns in the data frame for the `df_cue` argument
+that need to be passed along to the final output data frame. Default: NULL.}
 
-\item{spt}{An integer indicates the number of graphs to split the CO2 time series into for MANUALLY selecting the time cues. Cutting CO2 time series into several sections will aid the visualization and selection of the time cues when the data is too long. Default: 1. This argument is used only when no data is provided for the "Time_keys" argument.}
+\item{df_Ta}{A data frame contains a column "Ta" with the air temperature values (ideally, this is temperature measured inside
+ of the chamber during the flux measurement; unit: degree C. This can be the same data frame as in `df_cue`.
+ See example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata). The length of Ta within the file must be the same as the number of flux measurements.
+Default: NULL, then the temperature used is either the average ambient air temperature measured by the LGR analyzer
+(column "AmbT_C") or, if the data measured by other analyzers, Ta input from function `LoadOther()`.}
 
[…]
