### 2019-10-22 c329a6e
updated the help files
