### 2019-10-22 7835abb
added rmd readme and contributing.md
