+++ README.md
@@ -1,46 +1,91 @@
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
 # FluxCalR
-This R package is used to calculate CO<sub>2</sub> and CH<sub>4</sub> gas fluxes measured with static chamber method. It provides an 
-    easy way to calculate multiple flux measurements from one data file. 
-    
+
+<!-- badges: start -->
+
+This R package is used to calculate CO<sub>2</sub> and CH<sub>4</sub>
+gas fluxes measured with static chamber method. It provides an easy way
+to calculate multiple flux measurements from one data file.
+<!-- badges: end -->
+
+The goal of FluxCalR is to …
+
+## Installation
+
 **Package installation**
 
-First, make sure the package `devtools` is installed in R. If not, install the package by: 
-```R
-install.packages("devtools")
+First, make sure the package `remotes` is installed in R. If not,
+install the package by:
+
+``` r
+install.packages("remotes")
 ```
+
 Then, install the `FluxCalR` package in R by:
-```R
-devtools::install_github("junbinzhao/FluxCalR")
+
+``` r
+remotes::install_github("junbinzhao/FluxCalR")
 ```
-The functioning of the package is based on other R packages: `tidyverse`, `lubridate`, `assertthat`, `magrittr` and they must be installed before using the functions in the `FluxCalR` package.
+
+The functioning of the package is based on other R packages: `dplyr`,
+`lubridate`, `assertthat`, `magrittr` and they must be installed before
+using the functions in the `FluxCalR` package.
 
 **Load data**
 
-The format of data input are specialized to that from **Los Gatos Research (LGR) Ultraportable Gas Analyzers** 
-    <http://www.lgrinc.com/analyzers/ultraportable/>. Use the the function **`Load_LGR()`** to load raw data from LGR; 
-    it automatically removes the extra lines at the beginning and the end, and convert timestamps into the format that is readable in R. See details and examples ``?Load_LGR()``.
-    
-A function (**`Load_other()`**) is provided to load and convert data from other sources (e.g., LICOR) into the format that is 
-    compliant with the flux calculation fuction. See details and examples ``?Load_other()``.
+The format of data input are specialized to that from **Los Gatos
+Research (LGR) Ultraportable Gas Analyzers**
+<http://www.lgrinc.com/analyzers/ultraportable/>. Use the the function
+**`Load_LGR()`** to load raw data from LGR; it automatically removes the
+extra lines at the beginning and the end, and convert timestamps into
+the format that is readable in R. See details and examples `?LoadLGR()`.
+
+A function (**`Load_other()`**) is provided to load and convert data
+from other sources (e.g., LICOR) into the format that is compliant with
+the flux calculation fuction. See details and examples `?LoadOther()`.
 
 **Calculate fluxes**
 
-The function **`FluxCal()`** calculates CO<sub>2</sub> and/or CH<sub>4</sub> flux rates based on the time cues provided for each measurement (i.e., either 
-    start or end time). Two options are available to input the time cues: 
-1. (default) after executing the function, manually clicking on a pop-up graph with CO<sub>2</sub> concentration time series to choose 
-    the END time, which can usually be recognized as the local "peaks" or "valleys" in the time series; or 
-2. loading a file (.csv) into the argument "Time_keys" with times (HH:MM:SS) indicating the start or end of each flux measurement. 
-    The header for the time must be either "Start" or "End". 
-    (see example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata)
-    
[…]
