+++ CONTRIBUTING.md
@@ -1,0 +1,3 @@
+# Contributing
+
+*To report problems, seek support or contribute, please contact the author Junbin Zhao (junbinzhao1985@gmail.com). Requests/suggestions for new features are also welcome.*
