+++ vignettes/FluxCalR-vignette.Rmd
@@ -1,0 +1,21 @@
+---
+title: "FluxCalR package"
+author: "Junbin Zhao (junbinzhao1985@gmail.com)"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{FluxCalR-vignette}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(FluxCalR)
+```
