### 2019-10-23 e6e9069
added the error when ext <1. updated the vignettes
