### 2019-10-23 29dea60
change the example and also test for window length < end-start
