### 2019-10-23 fc65dd7
edit the error message for start or end
