### 2019-10-23 918f0d8
update help and debuged
