+++ man/FluxCal.Rd
@@ -77,6 +77,7 @@ Based on the time cues and window width provided for the calculation, the functi
  drawn for checkup purposes.
 }
 \examples{
+library(FluxCalR)
 #### data from LGR
 # get the directory of the example LGR raw data
 example_data1 <- system.file("extdata", "Flux_example_1_LGR.txt", package = "FluxCalR")
@@ -92,7 +93,7 @@ time_cue <- SelCue(Flux_lgr,flux = "CO2",cue = "End",save = F)
 # calculate the fluxes over a 3-minute window using the manually selected cues
 Flux_output1 <- FluxCal(data = Flux_lgr,
                         win = 3,
-                        vol = 208,Area = 0.26,
+                        vol = 208, area = 0.26,
                         df_cue = time_cue,
                         cue_type = "End",
                         output = FALSE) # don't create a output file
@@ -103,7 +104,7 @@ Example_cue1 <- system.file("extdata", "Time & Ta_1.csv", package = "FluxCalR") 
 Time_Ta1 <- read.csv(Example_cue1)
 Flux_output2 <- FluxCal(data = Flux_lgr,
                         win = 3,
-                        vol = 208, Area = 0.26,
+                        vol = 208, area = 0.26,
                         df_cue = Time_Ta1,
                         cue_type = "Start_End", # use both start and end time of each measurement as cues
                         other = c("Plot","Light_Dark"), # also pass other columns into the final output
@@ -128,10 +129,10 @@ Time_Ta2 <- read.csv(Example_cue2)
 Flux_output3 <- FluxCal(data = Flux_other,
                         cal = "CO2", # only calculate CO2 flux
                         win = 3,
-                        vol = 208, Area = 0.26,
+                        vol = 208, area = 0.26,
                         df_cue = Time_Ta2,
                         cue_type = "Start",
-                        other = c("Plot","Light_Dark")
+                        other = c("Plot","Light_Dark"),
                         output = FALSE) # don't create a output file
 Flux_output3
 
