+++ man/LoadOther.Rd
@@ -32,7 +32,7 @@ value can be used. Default: 25 (unit: degree C).}
 
 \item{sep}{A string indicates delimiter that separates columns. Default: ",".}
 
-\item{...}{Other arguments used in the function `read.table()`. See details `?read.table()`.}
+\item{...}{Other arguments used in the function \code{\link[base]{read.table}} `read.table()`. See details `?read.table()`.}
 }
 \value{
 A dataframe with all data in the right format for functions `SelCue()` and `FluxCal()`.
