+++ DESCRIPTION
@@ -12,6 +12,8 @@ Encoding: UTF-8
 LazyData: true
 Imports: dplyr, lubridate, assertthat, magrittr
 RoxygenNote: 6.1.1
-Suggests: knitr,
-    rmarkdown
+Suggests: 
+    knitr,
+    rmarkdown,
+    testthat
 VignetteBuilder: knitr
