+++ man/LoadOther.Rd
@@ -16,14 +16,14 @@ LoadOther(file, time, time_format = "mdy_HMS", date_ms = NULL,
 Other format, "dmy_HMS" or "ymd_HMS", can also be taken, depending on the data timestamp from the analyzer.
 In case that the date and time are recorded separately, "HMS" can be used here and add the date to the "date_ms" argument.}
 
-\item{date_ms}{A string indicates date of the measurements in the format of *"mm/dd/yyyy"*.
+\item{date_ms}{A string indicates date of the measurements in the format of \emph{"mm/dd/yyyy"}.
 This argument is used only when "HMS" has assigned to the argument `time_format`.}
 
 \item{CO2}{A string indicates column name of the CO2 concentrations in the data file. Default: no CO2 is measured.
-Note that at least one of `CO2` and `CH4` has to be specified.}
+Note that at least one of \code{CO2} and `CH4` argument has to be specified.}
 
 \item{CH4}{A string indicates column name of the CH4 concentrations in the data file. Default: no CH4 is measured.
-Note that at least one of `CO2` and `CH4` has to be specified.}
+Note that at least one of `CO2` and `CH4` argument has to be specified.}
 
 \item{Ta}{A string indicates column name of ambient air temperature in the data file. If not measured, a constant
 value can be used. Default: 25 (unit: degree C).}
@@ -32,7 +32,7 @@ value can be used. Default: 25 (unit: degree C).}
 
 \item{sep}{A string indicates delimiter that separates columns. Default: ",".}
 
-\item{...}{Other arguments used in the function \code{\link[base]{read.table}} `read.table()`. See details `?read.table()`.}
+\item{...}{Other arguments used in the function \code{\link[utils]{read.table()}}.}
 }
 \value{
 A dataframe with all data in the right format for functions `SelCue()` and `FluxCal()`.
