+++ R/FluxCal.R
@@ -1,54 +1,56 @@
 #' @title Calculate CO2 and CH4 gas fluxes
 #'
-#' @description A function to calculate CO2 and CH4 gas fluxes from the data loaded by the function `LoadLGR()` or `Load_other()`.
-#' It takes a time cue data frame (argument `df_cue`), either created by the function `SelCue()` or prepared by the user following
-#' the format of "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata,
+#' @description A function to calculate CO2 and CH4 gas fluxes from the data loaded by the function \code{\link{LoadLGR}}
+#' or \code{\link{LoadOther}}.
+#' It takes a time cue data frame (argument \code{df_cue}), either created by the function \code{\link{SelCue}} or prepared by the
+#' user following the format of
+#'  \link[=https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata]{"Time & Ta_#.csv"}
 #' to separate the measurements and then calculate the fluxes for all the measurements at once.
-#' Note that the header for the time cue column must be either *"Start"* or *"End"*.
+#' Note that the header for the time cue column must be either \strong{"Start"} or \strong{"End"}.
 #' Based on the time cues and window width provided for the calculation, the function will automatically scan over data that cover
 #'  1.5x (default) length of the window and calculate the fluxes based on the best linear regression (i.e., largest R2).
 #'  After the calculations are done, a graph with regression lines plotted on the CO2 and/or CH4 concentration time series can be
 #'  drawn for checkup purposes.
 #'
-#' @param data A dataframe generated by the function `LoadLGR` or `LoadOther`.
+#' @param data A dataframe generated by the function \code{\link{LoadLGR}} or \code{\link{LoadOther}}.
 #' @param cal A string, either "CO2_CH4" (default),"CO2" or "CH4", indicates which gas flux it is calculated for.
 #' @param win A number indicates the window width for the flux calculation, unit: minutes.
 #' @param vol A number indicates volume of the chamber; unit: dm^3 or L.
 #' @param area A number indicates base area of the chamber; unit: m^2.
 #' @param df_cue A data frame that includes "Start" and/or "End" time (HH:MM:SS) of each measurement.
-#' The header for the time must be *"Start"* or *"End"*. This data frame can either be created by the function `SelCue()`
-#' or be prepared by the user
+#' The header for the time must be either \strong{"Start"} or \strong{"End"}. This data frame can either be created by the
+#' function \code{\link{SelCue}} or be prepared by the user
 #' (see example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
 #' @param cue_type A string, either "Start", "End" (default) or "Start_End", indicates if start, end or both time in the
-#' data frame assigned to `df_cue` will be used as the cues . When "Start_End" is chosen, both "Start" and "End" columns have to
-#' be present in the data frame for `df_cue` and the flux with the largest R2 within the range will be calculated without
-#' considering the argument `ext`.
-#' @param other A vector of strings indicates the names of other columns in the data frame for the `df_cue` argument
+#' data frame assigned to \code{df_cue} will be used as the cues . When "Start_End" is chosen, both "Start" and "End" columns
+#' have to be present in the data frame for \code{df_cue} and the flux with the largest R2 within the range will be calculated
+#' without considering the argument \code{ext}.
+#' @param other A vector of strings indicates the names of other columns in the data frame for the \code{df_cue} argument
 #' that need to be passed along to the final output data frame. Default: NULL.
[…]
