+++ man/LoadOther.Rd
@@ -17,13 +17,13 @@ Other format, "dmy_HMS" or "ymd_HMS", can also be taken, depending on the data t
 In case that the date and time are recorded separately, "HMS" can be used here and add the date to the "date_ms" argument.}
 
 \item{date_ms}{A string indicates date of the measurements in the format of \emph{"mm/dd/yyyy"}.
-This argument is used only when "HMS" has assigned to the argument `time_format`.}
+This argument is used only when "HMS" has assigned to the argument \code{time_format}.}
 
 \item{CO2}{A string indicates column name of the CO2 concentrations in the data file. Default: no CO2 is measured.
-Note that at least one of \code{CO2} and `CH4` argument has to be specified.}
+Note that at least one of \code{CO2} and \code{CH4} argument has to be specified.}
 
 \item{CH4}{A string indicates column name of the CH4 concentrations in the data file. Default: no CH4 is measured.
-Note that at least one of `CO2` and `CH4` argument has to be specified.}
+Note that at least one of \code{CO2} and \code{CH4} argument has to be specified.}
 
 \item{Ta}{A string indicates column name of ambient air temperature in the data file. If not measured, a constant
 value can be used. Default: 25 (unit: degree C).}
@@ -32,14 +32,14 @@ value can be used. Default: 25 (unit: degree C).}
 
 \item{sep}{A string indicates delimiter that separates columns. Default: ",".}
 
-\item{...}{Other arguments used in the function \code{\link[utils]{read.table()}}.}
+\item{...}{Other arguments used in the function \code{\link[utils]{read.table}}.}
 }
 \value{
-A dataframe with all data in the right format for functions `SelCue()` and `FluxCal()`.
+A dataframe with all data in the right format for functions \code{\link{SelCue}} and \code{\link{FluxCal}}.
 }
 \description{
 Function used to load raw data output from other similar analyzers (e.g. LICOR-7810) to be called
-in the functions `SelCue()` and `FluxCal()`. This function extract the data (i.e., CO2 and/or CH4 concentration and ambient
+in the functions \code{\link{SelCue}} and \code{\link{FluxCal}}. This function extract the data (i.e., CO2 and/or CH4 concentration and ambient
 temperature) required to calculate the fluxes and converts the timestamps into POSIXct for further processing.
 }
 \examples{
