+++ man/FluxCal.Rd
@@ -10,7 +10,7 @@ FluxCal(data, cal = "CO2_CH4", win, vol, area, df_cue,
   ylim_CO2 = NULL, ylim_CH4 = NULL)
 }
 \arguments{
-\item{data}{A dataframe generated by the function `LoadLGR` or `LoadOther`.}
+\item{data}{A dataframe generated by the function \code{\link{LoadLGR}} or \code{\link{LoadOther}}.}
 
 \item{cal}{A string, either "CO2_CH4" (default),"CO2" or "CH4", indicates which gas flux it is calculated for.}
 
@@ -21,27 +21,27 @@ FluxCal(data, cal = "CO2_CH4", win, vol, area, df_cue,
 \item{area}{A number indicates base area of the chamber; unit: m^2.}
 
 \item{df_cue}{A data frame that includes "Start" and/or "End" time (HH:MM:SS) of each measurement.
-The header for the time must be *"Start"* or *"End"*. This data frame can either be created by the function `SelCue()`
-or be prepared by the user
+The header for the time must be either \strong{"Start"} or \strong{"End"}. This data frame can either be created by the
+function \code{\link{SelCue}} or be prepared by the user
 (see example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).}
 
 \item{cue_type}{A string, either "Start", "End" (default) or "Start_End", indicates if start, end or both time in the
-data frame assigned to `df_cue` will be used as the cues . When "Start_End" is chosen, both "Start" and "End" columns have to
-be present in the data frame for `df_cue` and the flux with the largest R2 within the range will be calculated without
-considering the argument `ext`.}
+data frame assigned to \code{df_cue} will be used as the cues . When "Start_End" is chosen, both "Start" and "End" columns
+have to be present in the data frame for \code{df_cue} and the flux with the largest R2 within the range will be calculated
+without considering the argument \code{ext}.}
 
-\item{other}{A vector of strings indicates the names of other columns in the data frame for the `df_cue` argument
+\item{other}{A vector of strings indicates the names of other columns in the data frame for the \code{df_cue} argument
 that need to be passed along to the final output data frame. Default: NULL.}
 
 \item{df_Ta}{A data frame contains a column "Ta" with the air temperature values (ideally, this is temperature measured inside
- of the chamber during the flux measurement; unit: degree C. This can be the same data frame as in `df_cue`.
+ of the chamber during the flux measurement; unit: degree C. This can be the same data frame as in \code{df_cue}.
  See example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
  Note the row number of the data frame must be the same as the number of flux measurements.
 Default: NULL, then the temperature used is either the average ambient air temperature measured by the LGR analyzer
-(column "AmbT_C") or, if the data measured by other analyzers, Ta input from function `LoadOther()`.}
+(column "AmbT_C") or, if the data measured by other analyzers, Ta input from function \code{\link{LoadOther}}.}
 
-\item{ext}{A number indicates a range of how many times of the window width (`win`) should the calculation scan through to
-choose the regression with the largest R2. Default: 1.5. This argument is ignored when `cue_type` is "Start_End".}
+\item{ext}{A number indicates a range of how many times of the window width (\code{win}) should the calculation scan through to
+choose the regression with the largest R2. Default: 1.5. This argument is ignored when \code{df_cue} is "Start_End".}
 
 \item{output}{A string includes output directory and file name (.csv) to export the calculated fluxes.
[…]
