+++ man/SelCue.Rd
@@ -8,7 +8,7 @@ SelCue(data, flux = "CO2", cue = "End", spt = 1, ylim = NULL,
   save = "Time_cue.csv")
 }
 \arguments{
-\item{data}{A dataframe generated by the function `LoadLGR()` or `LoadOther()`.}
+\item{data}{A dataframe generated by the function \code{\link{LoadLGR}} or \code{\link{LoadOther}}.}
 
 \item{flux}{A string, either "CO2" (default) or "CH4", indicates that either CO2 or CH4 concentration
 should be plotted for selecting the time cues.}
@@ -37,7 +37,7 @@ A data frame that includes two columns:
 \description{
 This is a function to manually select and extract the start or the end time of each measurement as time cues by
 clicking the corresponding start or end measurement points on the interactive graphs generated by the function.
-The data frame returned from this function can be assigned to the argument `df_cue` in the `FluxCal()` function.
+The data frame returned from this function can be assigned to the argument \code{df_cue} in the \code{\link{FluxCal}} function.
 Note that the time cue data frame can also be prepared by the user independent of this function following the format of
 "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
 }
