### 2019-11-01 391240d
added example graphs and output time cue data
