### 2019-11-01 786cbab
moved the images for vignette to the folder with vignette
