### 2019-11-01 369b79b
reformat the vignette
