+++ README.md
@@ -5,12 +5,59 @@
 
 <!-- badges: start -->
 
-This R package is used to calculate CO<sub>2</sub> and CH<sub>4</sub>
-gas fluxes measured with static chamber method. It provides an easy way
-to calculate multiple flux measurements from one data file.
-<!-- badges: end -->
+This R package `FluxCalR` is used to calculate CO<sub>2</sub> and
+CH<sub>4</sub> gas fluxes measured with static chamber method. It
+provides an easy way to calculate multiple flux measurements from one
+data file. The package directly takes the raw data exported by a gas
+analyzer (e.g. LGR Ultraportable Gas Analyzers) after chamber
+measurements and calculate fluxes of multiple measurements at once from
+the data file. Specifically, users can use the function `LoadLGR()` or
+`LoadOther()` to load the raw data file that includes the gas
+concentration records. Then, create a separate data frame that includes
+the time cues that each measurement starts and/or ends corresponding to
+the time in the raw data file. Based on the time cues, the fluxes of all
+the measurements can be calculated and exported using the function
+`FluxCal()`. The package has several features that facilitate the
+process of flux calculation:
 
-The goal of FluxCalR is to …
+  - **An interactive way of selecting the measurement time cues** - The
+    package includes the function `SelCue()` that one can use to select
+    the time cues (start or end of each measurement) manually by
+    clicking on an interactive graph of the entire gas concentration
+    time series from the raw data file. The data frame returned from the
+    function can be directly used as the argument `df_cue` of
+    `FluxCal()` for the flux calculation. The time cue data frame can
+    also be saved (default) and used again to get reproducible results.
+
+  - **Flexible input** - Depending on the measurement protocol and
+    requirement of the users, the `FluxCal()` function is flexible in
+    the sense that 1) it takes either “Start”, “End” or both
+    “Start\_End” of each measurement (argument `cue_type`) as time
+    cues to identify each measurement; 2) it calculates “CO2”, “CH4” or
+    both “CO2\_CH4” fluxes at the same time (argument `cal`).
+
+  - **A dynamic measurement window** - The flux calculation
+    automatically scans through a range of 1.5x (default) of the
+    measurement window and the best linear regression (largest R2) is
+    used to calculate the final fluxes. This protocol optimize the
+    representative and reproducibility of the calculated flux. This
+    dynamic range can be changed in the argument `ext` of the function
+    `FluxCal()`.
+
+  - **A checkup graph after the calculation** - By default (argument
+    `check_plot`), a graph is plotted after the calculation with
+    regression lines plotted on the CO2 and/or CH4 concentration time
+    series for checkup purposes. This post-calculation feature will help
+    users identify any flux calculation that is based on the unwanted
+    section of the data.
+
+  - **Easy to be integrated into workflows** - The `FluxCal()` returns a
+    data frame with the calculated flux and identifier of each
+    measurement (e.g., plot ID, see argument `Other`) that can be
+    integrated into users’ data processing/analyzing workflow. At the
+    same time, it also saves the data frame to a “.csv” file that can be
+    loaded by R again later or used in other programs for non-R users.
+    <!-- badges: end -->
 
 ## Installation
 
@@ -28,62 +75,104 @@ remotes::install_github("junbinzhao/FluxCalR")
 ```
 
 The functioning of the package is based on other R packages: `dplyr`,
-`lubridate`, `assertthat`, `magrittr` and they must be installed before
[…]
