### 2019-11-04 e01405d
add installation with vignette
