+++ README.Rmd
@@ -35,9 +35,9 @@ First, make sure the package `remotes` is installed in R. If not, install the pa
 ```r
 install.packages("remotes")
 ```
-Then, install the `FluxCalR` package in R by:
+Then, install the `FluxCalR` package (with the vignette) in R by:
 ```r
-remotes::install_github("junbinzhao/FluxCalR")
+remotes::install_github("junbinzhao/FluxCalR",build_vignettes = TRUE)
 ```
 The functioning of the package is based on other R packages: `dplyr`, `lubridate`, `assertthat`, `magrittr` and they will/must be installed before using the functions in the `FluxCalR` package.
 
@@ -116,7 +116,7 @@ This is how the output data frame looks like and the last column ("Flux") includ
 head(Flux_output1)
 ```
 
-For more details of the functions and examples, please check the vignette (`browseVignettes(FluxCalR)`) and the help documents (`?LoadLGR()`,`?LoadOther()`,`?SelCue()`,`?FluxCal()`). 
+For more details of the functions and examples, please check the vignette (`browseVignettes("FluxCalR")`) and the help documents (`?LoadLGR()`,`?LoadOther()`,`?SelCue()`,`?FluxCal()`). 
 
 
 ### Please cite the package as:
