+++ README.Rmd
@@ -35,7 +35,7 @@ First, make sure the package `remotes` is installed in R. If not, install the pa
 ```r
 install.packages("remotes")
 ```
-Then, install the `FluxCalR` package (with the vignette) in R by:
+Then, install the `FluxCalR` package (with vignettes) in R by:
 ```r
 remotes::install_github("junbinzhao/FluxCalR",build_vignettes = TRUE)
 ```
