### 2019-11-04 57eed98
don't test the select cue example
