### 2019-11-04 945a080
add don't test to examples
