+++ R/FluxCal.R
@@ -53,6 +53,7 @@
 #' for checkup purposes.
 #'
 #' @examples
+#' \donttest{
 #' library(FluxCalR)
 #' #### data from LGR
 #' # get the directory of the example LGR raw data
@@ -64,7 +65,7 @@
 #'                     time_format = "mdy_HMS")
 #'
 #' # manually select the end of each measurement as time cues
-#' time_cue <- SelCue(Flux_lgr,flux = "CO2",cue = "End",save = F)
+#' time_cue <- SelCue(Flux_lgr,flux = "CO2",cue = "End",save = FALSE)
 #'
 #' # calculate the fluxes over a 3-minute window using the manually selected cues
 #' Flux_output1 <- FluxCal(data = Flux_lgr,
@@ -111,6 +112,7 @@
 #'                         other = c("Plot","Light_Dark"),
 #'                         output = FALSE) # don't create a output file
 #' Flux_output3
+#' }
 #'
 #' @export
 ## Flux calculation function ---------------
