+++ man/LoadOther.Rd
@@ -47,9 +47,9 @@ library(FluxCalR)
 # get the directory of the example data
 example_data <- system.file("extdata", "Flux_example_2_other.csv", package = "FluxCalR")
 # load the data
-Flux <- LoadOther(example_data,
-                  Time = "Date_time",
-                  Time_format = "mdy_HMS",
+Flux <- LoadOther(file = example_data,
+                  time = "Date_time",
+                  time_format = "mdy_HMS",
                   CO2 = "CO2_PPM",
                   Ta = "Tem_C")
 
