### 2019-11-04 680d232
added test for loading functions
