+++ R/Load.R
@@ -29,9 +29,10 @@ LoadLGR <- function(file,
   flux <- read.table(file, sep= ",", header = TRUE,skip = 1,fill = TRUE,stringsAsFactors = FALSE)
   flux <- subset(flux,!is.na(flux[,2])) # remove the comments generated by LGR
   # replace the time with "lubridated" time with different format
-  flux$Time <- try(lubridate::parse_date_time(flux$Time,orders = time_format),silent = TRUE)
+  flux$Time <- try(suppressWarnings(lubridate::parse_date_time(flux$Time,orders = time_format)),
+                   silent = TRUE)
   # error if time column is note correctly referred to
-  if (class(flux$Time)[1]=="try-error"){
+  if (class(flux$Time)[1]=="try-error" | is.na(flux$Time)[1]){
     stop("Error: check if the 'time_format' argument is correct!")
   }
   return(flux)
@@ -100,11 +101,13 @@ LoadOther <- function(file,
     Time <- try(paste0(Date_ms,"_",flux1[,time]),silent = TRUE) # if only time is provided, then combine the time with date first
     Time <- try(lubridate::mdy_hms(Time),silent = TRUE)
   } else { # if it is full date format
-    Time <- try(lubridate::parse_date_time(flux1[,time],orders = time_format),silent = TRUE)
+    Time <- suppressWarnings(try(lubridate::parse_date_time(flux1[,time],
+                                                            orders = time_format),
+                                 silent = TRUE))
   }
 
   # error if time column is note correctly referred to
-  if (class(Time)[1]=="try-error"){
+  if (class(Time)[1]=="try-error" | is.na(Time)[1]){
     stop("Error: check if the 'time', 'time_format' and/or 'date_ms' arguments are correct!")
   }
   flux <- data.frame(Time)
