### 2019-11-04 dee226c
add calcualation test
