### 2019-11-05 f8e2157
changed the tci badge and added the need of the package [ci skip]
