+++ paper.md
@@ -11,21 +11,45 @@ authors:
     orcid: 0000-0001-5142-4901
     affiliation: "1, 2" # (Multiple affiliations must be quoted)
 affiliations:
- - name: Department of Terrestrial Ecology, Division of Environment and Natural Resources, Norwegian Institute of Bioeconomy Research, Ås, Norway
+ - name: Department of Biogeochemistry and Soil Quality, Division of Environment and Natural Resources, Norwegian Institute of Bioeconomy Research, Ås, Norway
    index: 1
  - name: Department of Biological Sciences, Florida International University, Miami, FL, USA
    index: 2
-date: 13 August 2019
+date: 05 November 2019
 bibliography: paper.bib
 ---
 
+# Statement of need
+
+As the main drivers of climate change, greenhouse gas (e.g., CO<sub>2</sub> and CH<sub>4</sub>) emissions have been monitored intensively across the globe. The static chamber is one of the most commonly used approaches for measuring greenhouse gas flux from ecosystems (e.g., stem/soil respiration, CH<sub>4</sub> emission, etc.) because of its easy implementation, high accuracy and low cost [@Pumpanen:2004]. To perform the measurements, commercial infrared gas analyzers (IRGA) are usually used to measure the changes of greenhouse gas concentrations within a chamber that covers an area of interest (e.g., soil surface) over a centain period of time (usually several minutes). The flux rates are, then, calculated from the recorded gas concentraion changes based on linear regressions:
+
+$$
+F = vol/(R*Ta*area)*dG/dt
+$$
+
+where *vol* is the volumn of the chamber (l), *R* is universal gas constant (l atm K^-1^ mol^-1^), *Ta* is the ambient temperature (K), *area* is the chamber base area (m^2^), and *dG/dt* is the linear slope of the measured gas concentration change over time *t* (ppm s^-1^).
+
+Eventhough the method itself is straight forward to conduct, to calculate the fluxes from the recorded gas concentration data can be tedious considering that the original dataset usually includes a mix of valid data and unwanted records irregularly placed between each other. In addtion, since the calculated flux rate is largely determined by the slope of the regression, choosing a window that delivers a better regression is important to to achieve the highest representativeness of the fluxes. The R^2^ of the regression is usually used to determine the quality of a flux measurement [@Zhao:2016; @Zhao:2017]. This leads to another problem when one selects fixed valid data windows for calculations based on his/her own judgement that may actually not be representive for the flux measurement. that covers the because that the data that is valid for calculation  and the window location for calculating each flux is fixed. Therefore, a tool that can efficiently calculate the concentration changes into fluxes in a more efficient and objective fashion is needed. 
+
+To my knowledge, there are already some softwares that has the function to calculate the fluxes measured using static chamber method. However, they are either compatible to only one designated system (e.g., [Licor Trace Gas Analyzer Platform](https://www.licor.com/env/products/trace_gas/), [Flux Puppy](https://www.sciencedirect.com/science/article/pii/S0168192319301522)), unable to calculate fluxes for multiple gases (e.g., [Flux Puppy](https://www.sciencedirect.com/science/article/pii/S0168192319301522)), or hard to handle continously measured concentrations (e.g., R package ['flux'](https://cran.r-project.org/web/packages/flux/index.html)). Therefore, there is still a need for a free, flexible and easy-to-use software that is designed to calculate ecosystem gas fluxes from continously measured concentrations. 
+
 # Summary
 
[…]
