+++ paper.md
@@ -1,8 +1,11 @@
 ---
 title: 'FluxCalR: a R package for calculating CO2 and CH4 fluxes from static chambers'
 tags:
+  - CO2
+  - CH4
   - greenhouse gas
-  - chamber method
+  - static chamber
+  - continuous measurement
   - Los Gatos Research 
   - Soil respiration
   - Ecosystem flux
@@ -19,38 +22,25 @@ date: 05 November 2019
 bibliography: paper.bib
 ---
 
-# Statement of need
+# Summary
 
-As the main drivers of climate change, greenhouse gas (e.g., CO<sub>2</sub> and CH<sub>4</sub>) emissions have been monitored intensively across the globe. The static chamber is one of the most commonly used approaches for measuring greenhouse gas flux from ecosystems (e.g., stem/soil respiration, CH<sub>4</sub> emission, etc.) because of its easy implementation, high accuracy and low cost [@Pumpanen:2004]. To perform the measurements, commercial infrared gas analyzers (IRGA) are usually used to measure the changes of greenhouse gas concentrations within a chamber that covers an area of interest (e.g., soil surface) over a centain period of time (usually several minutes). The flux rates are, then, calculated from the recorded gas concentraion changes based on linear regressions:
+As the main drivers of climate change, greenhouse gas (e.g., CO~2~ and CH<sub>4</sub>) emissions have been monitored intensively across the globe. The static chamber is one of the most commonly used approaches for measuring greenhouse gas flux from ecosystems (e.g., stem/soil respiration, CH<sub>4</sub> emission, etc.) because of its easy implementation, high accuracy and low cost [@Pumpanen:2004]. To perform the measurements, an infrared gas analyzers (IRGA) are usually used to measure the changes of greenhouse gas concentrations within a closed chamber that covers an area of interest (e.g., soil surface) over a certain period of time (usually several minutes). The flux rates are, then, calculated from the recorded gas concentrations assuming that the changing rate is linear:
 
 $$
 F = vol/(R*Ta*area)*dG/dt
 $$
 
-where *vol* is the volumn of the chamber (l), *R* is universal gas constant (l atm K^-1^ mol^-1^), *Ta* is the ambient temperature (K), *area* is the chamber base area (m^2^), and *dG/dt* is the linear slope of the measured gas concentration change over time *t* (ppm s^-1^).
+where *vol* is the volume of the chamber (l), *R* is universal gas constant (l atm K^-1^ mol^-1^), *Ta* is the ambient temperature (K), *area* is the area of the chamber base (m^2^), and *dG/dt* is the rate (i.e., the slope of the linear regression) of the measured gas concentration change over time *t* (ppm s^-1^).
 
-Eventhough the method itself is straight forward to conduct, to calculate the fluxes from the recorded gas concentration data can be tedious considering that the original dataset usually includes a mix of valid data and unwanted records irregularly placed between each other. In addtion, since the calculated flux rate is largely determined by the slope of the regression, choosing a window that delivers a better regression is important to to achieve the highest representativeness of the fluxes. The R^2^ of the regression is usually used to determine the quality of a flux measurement [@Zhao:2016; @Zhao:2017]. This leads to another problem when one selects fixed valid data windows for calculations based on his/her own judgement that may actually not be representive for the flux measurement. that covers the because that the data that is valid for calculation  and the window location for calculating each flux is fixed. Therefore, a tool that can efficiently calculate the concentration changes into fluxes in a more efficient and objective fashion is needed. 
[…]
