+++ paper.bib
@@ -37,3 +37,24 @@ title = {{Long-term enhanced winter soil frost alters growing season CO2 fluxes 
 url = {http://doi.wiley.com/10.1111/gcb.13621},
 year = {2017}
 }
+
+@article{Jurasinski:2014,
+    title = {flux: Flux rate calculation from dynamic closed chamber measurements},
+    author = {Gerald Jurasinski and Franziska Koebsch and Anke Guenther and Sascha Beetz},
+    year = {2014},
+    note = {R package version 0.3-0},
+    url = {https://CRAN.R-project.org/package=flux},
+  }
+
+@article{Carbone:2019,
+title = "Flux Puppy – An open-source software application and portable system design for low-cost manual measurements of CO2 and H2O fluxes",
+journal = "Agricultural and Forest Meteorology",
+volume = "274",
+pages = "1 - 6",
+year = "2019",
+issn = "0168-1923",
+doi = "https://doi.org/10.1016/j.agrformet.2019.04.012",
+url = "http://www.sciencedirect.com/science/article/pii/S0168192319301522",
+author = "Mariah S. Carbone and Bijan Seyednasrollah and Tim T. Rademacher and David Basler and James M. Le Moine and Samuel Beals and James Beasley and Andrew Greene and Joseph Kelroy and Andrew D. Richardson",
+keywords = "Manual chamber CO measurements, LI-820, LI-830, LI-840, LI-850, Open source, Portable system, Respiration",
+}
