+++ paper.md
@@ -1,8 +1,8 @@
 ---
-title: 'FluxCalR: a R package for calculating CO2 and CH4 fluxes from static chambers'
+title: 'FluxCalR: a R package for calculating CO~2~ and CH~4~ fluxes from static chambers'
 tags:
-  - CO2
-  - CH4
+  - CO~2~
+  - CH~4~
   - greenhouse gas
   - static chamber
   - continuous measurement
@@ -24,19 +24,19 @@ bibliography: paper.bib
 
 # Summary
 
-As the main drivers of climate change, greenhouse gas (e.g., CO~2~ and CH<sub>4</sub>) emissions have been monitored intensively across the globe. The static chamber is one of the most commonly used approaches for measuring greenhouse gas flux from ecosystems (e.g., stem/soil respiration, CH<sub>4</sub> emission, etc.) because of its easy implementation, high accuracy and low cost [@Pumpanen:2004]. To perform the measurements, an infrared gas analyzers (IRGA) are usually used to measure the changes of greenhouse gas concentrations within a closed chamber that covers an area of interest (e.g., soil surface) over a certain period of time (usually several minutes). The flux rates are, then, calculated from the recorded gas concentrations assuming that the changing rate is linear:
+As the main drivers of climate change, greenhouse gas (e.g., CO~2~ and CH~4~) emissions have been monitored intensively across the globe. The static chamber is one of the most commonly used approaches for measuring greenhouse gas flux from ecosystems (e.g., stem/soil respiration, CH~4~ emission, etc.) because of its easy implementation, high accuracy and low cost [@Pumpanen:2004]. To perform the measurements, a gas analyzer is usually used to measure the changes of greenhouse gas concentrations within a closed chamber that covers an area of interest (e.g., soil surface) over a certain period of time (usually several minutes). The flux rates are, then, calculated from the recorded gas concentrations assuming that the changing rate is linear:
 
 $$
-F = vol/(R*Ta*area)*dG/dt
+F = vol/(R∙Ta∙area)∙dG/dt
 $$
 
 where *vol* is the volume of the chamber (l), *R* is universal gas constant (l atm K^-1^ mol^-1^), *Ta* is the ambient temperature (K), *area* is the area of the chamber base (m^2^), and *dG/dt* is the rate (i.e., the slope of the linear regression) of the measured gas concentration change over time *t* (ppm s^-1^).
 
 Even though the method itself appears to be straight forward, to calculate the fluxes from the raw gas concentration data can be tedious, which includes extracting the data for each measurement from an irregular mix of valid data and unwanted records and choosing a window that delivers the best regression for each measurement to ensure the quality and representativeness of the calculated flux. The R^2^ of the regression is usually taken to determine the quality of a flux measurement and a flux with low R^2^ is usually discarded [@Zhao:2016; @Zhao:2017]. Therefore, a tool that can calculate the gas concentration changes into fluxes in an efficient and objective fashion is needed. 
 
-There are some software that calculate the fluxes measured using static chamber method. However, they are either compatible to only one designated system (e.g., [Li-Cor Trace Gas Analyzer Platform](https://www.licor.com/env/products/trace_gas/), [Flux Puppy](https://www.sciencedirect.com/science/article/pii/S0168192319301522)), unable to calculate fluxes for multiple gases (e.g., [Flux Puppy](https://www.sciencedirect.com/science/article/pii/S0168192319301522)), or hard to handle continuously measured gas concentrations (e.g., R package ['flux'](https://cran.r-project.org/web/packages/flux/index.html)). Therefore, there is still a need for a free, flexible and easy-to-use software that is designed to calculate ecosystem gas fluxes from continuously measured gas concentrations. 
[…]
