### 2019-11-06 c24819f
update equation [ci skip]
