+++ paper.bib
@@ -1,5 +1,5 @@
 @article{Pumpanen:2004,
-title = "Comparison of different chamber techniques for measuring soil CO2 efflux",
+title = "Comparison of different chamber techniques for measuring soil CO~2~ efflux",
 journal = "Agricultural and Forest Meteorology",
 volume = "123",
 number = "3",
@@ -33,7 +33,7 @@ doi = {10.1111/gcb.13621},
 issn = {13541013},
 journal = {Global Change Biology},
 pages = {1-15},
-title = {{Long-term enhanced winter soil frost alters growing season CO2 fluxes through its impact on vegetation development in a boreal peatland}},
+title = {{Long-term enhanced winter soil frost alters growing season CO~2~ fluxes through its impact on vegetation development in a boreal peatland}},
 url = {http://doi.wiley.com/10.1111/gcb.13621},
 year = {2017}
 }
@@ -47,7 +47,7 @@ year = {2017}
   }
 
 @article{Carbone:2019,
-title = "Flux Puppy – An open-source software application and portable system design for low-cost manual measurements of CO2 and H2O fluxes",
+title = "Flux Puppy – An open-source software application and portable system design for low-cost manual measurements of CO~2~ and H~2~O fluxes",
 journal = "Agricultural and Forest Meteorology",
 volume = "274",
 pages = "1 - 6",
