+++ paper.bib
@@ -1,5 +1,5 @@
 @article{Pumpanen:2004,
-title = "Comparison of different chamber techniques for measuring soil CO<sub>2<sub> efflux",
+title = "Comparison of different chamber techniques for measuring soil CO<sub>2<\sub> efflux",
 journal = "Agricultural and Forest Meteorology",
 volume = "123",
 number = "3",
