### 2019-11-06 48bbf5d
updated read me and help.[ci skip]
