+++ R/FluxCal.R
@@ -13,10 +13,10 @@
 #'  drawn for checkup purposes.
 #'
 #' @param data A dataframe generated by the function \code{\link{LoadLGR}} or \code{\link{LoadOther}}.
-#' @param cal A string, either "CO2_CH4" (default),"CO2" or "CH4", indicates which gas flux it is calculated for.
 #' @param win A number indicates the window width for the flux calculation, unit: minutes.
 #' @param vol A number indicates volume of the chamber; unit: dm^3 or l.
 #' @param area A number indicates base area of the chamber; unit: m^2.
+#' @param cal A string, either "CO2_CH4" (default),"CO2" or "CH4", indicates which gas flux it is calculated for.
 #' @param df_cue A data frame that includes "Start" and/or "End" time (HH:MM:SS) of each measurement.
 #' The header for the time must be either \strong{"Start"} or \strong{"End"}. This data frame can either be created by the
 #' function \code{\link{SelCue}} or be prepared by the user
@@ -117,10 +117,10 @@
 #' @export
 ## Flux calculation function ---------------
 FluxCal <- function(data,
-                    cal = "CO2_CH4",
                     win,
                     vol,
                     area,
+                    cal = "CO2_CH4",
                     df_cue,
                     cue_type = "End",
                     other = NULL,
