+++ man/FluxCal.Rd
@@ -4,7 +4,7 @@
 \alias{FluxCal}
 \title{Calculate CO2 and CH4 gas fluxes}
 \usage{
-FluxCal(data, cal = "CO2_CH4", win, vol, area, df_cue,
+FluxCal(data, win, vol, area, cal = "CO2_CH4", df_cue,
   cue_type = "End", other = NULL, df_Ta = NULL, ext = 1.5,
   output = "Flux_output.csv", digits = 3, check_plot = TRUE,
   ylim_CO2 = NULL, ylim_CH4 = NULL)
@@ -12,13 +12,13 @@ FluxCal(data, cal = "CO2_CH4", win, vol, area, df_cue,
 \arguments{
 \item{data}{A dataframe generated by the function \code{\link{LoadLGR}} or \code{\link{LoadOther}}.}
 
-\item{cal}{A string, either "CO2_CH4" (default),"CO2" or "CH4", indicates which gas flux it is calculated for.}
-
 \item{win}{A number indicates the window width for the flux calculation, unit: minutes.}
 
 \item{vol}{A number indicates volume of the chamber; unit: dm^3 or l.}
 
 \item{area}{A number indicates base area of the chamber; unit: m^2.}
+
+\item{cal}{A string, either "CO2_CH4" (default),"CO2" or "CH4", indicates which gas flux it is calculated for.}
 
 \item{df_cue}{A data frame that includes "Start" and/or "End" time (HH:MM:SS) of each measurement.
 The header for the time must be either \strong{"Start"} or \strong{"End"}. This data frame can either be created by the
