+++ README.Rmd
@@ -21,7 +21,7 @@ knitr::opts_chunk$set(
 
 ### Why is the package needed?
 
-The static close chamber technique is a widely used approach for measuring greenhouse gas fluxes from different sources in an ecosystem (e.g. soil respiration). While the method itself is easy and cheap to implement, the post-calculation of fluxes from the measured concentrations is a cumbersome process. Previous softwares that can calculate the fluxes were either compatible to only one designated system (e.g., [Licor Trace Gas Analyzer Platform](https://www.licor.com/env/products/trace_gas/), [Flux Puppy](https://www.sciencedirect.com/science/article/pii/S0168192319301522)), unable to calculate fluxes for multiple gases (e.g., [Flux Puppy](https://www.sciencedirect.com/science/article/pii/S0168192319301522)), or hard to handle continously measured concentrations (e.g., R package ['flux'](https://cran.r-project.org/web/packages/flux/index.html)). 
+The static close chamber technique is a widely used approach for measuring greenhouse gas fluxes from different sources in an ecosystem (e.g. soil respiration). While the method itself is easy and cheap to implement, the post-calculation of fluxes from the measured concentrations is a cumbersome process. Previous softwares that can calculate the fluxes were either compatible to only one designated system (e.g., [Licor Trace Gas Analyzer Platform](https://www.licor.com/env/products/trace_gas/), [Flux Puppy](https://www.sciencedirect.com/science/article/pii/S0168192319301522)), unable to calculate fluxes for multiple gases (e.g., [Flux Puppy](https://www.sciencedirect.com/science/article/pii/S0168192319301522)), or hard to handle continuously measured concentrations (e.g., R package ['flux'](https://cran.r-project.org/web/packages/flux/index.html)). 
 
 ### About the package
 
@@ -92,6 +92,11 @@ knitr::include_graphics("vignettes/images/timecue.png") # add time cue graph wit
 time_cue <- read.csv("vignettes/Time_cue.csv")
 ```
 
+This is how the "time_cue" data frame looks like. Instead of having it created by the function `SelCue()`, one can also manually create such a file and load it into R as a data frame. But the column names ("End" or "Start") and time format must be in line with the "time_cue" here, or see example files "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).  
+
+```{r}
+time_cue
+```
 
 After having the time cues saved as a data frame 'time_cue', we can do the calculations now using the function `FluxCal()`! In the function, there are **5 arguments** that have to be assigned by the users without default: `data` (the data frame we got from `LoadLGR()`), `win` (the window size flux calculation, unit: minute),`vol` (chamber volumn in L),`area` (chamber base area in m^2) and `df_cue` (the data frame we got from `SelCue()`). After the function is executed with the arguments assigned, the calculation will be done and a check-up graph with all the regression lines drawn on top of the gas concentration time series will pop up by default for the user to visually check if there is any mistake occured to the calcualtions. 
 
