### 2019-11-07 1327d67
proof read the paper and readme files
