### 2019-11-07 #1751 @junbinzhao
@arfon @jhollist @ha0ye Thank you so much for the reviews! They are indeed very helpful and practical! I have revised the package as well as the paper accordingly. I believe they are now much improved. I have added the point-to-point responses under [@jhollist review](https://github.com/junbinzhao/FluxCalR/issues/2) and [@ha0ye review](https://github.com/junbinzhao/FluxCalR/issues/3). Please let me know your further comments!

Thank you very much again!!

Junbin
