### 2019-11-07 #1751 @junbinzhao
> > @arfon I have finished the revision and all the updates are done in the github now. but I don't know how to submit the point-to-point responses to the reviews. Do I attach the file here directly? Thank you!
>
> Ideally it would be good to respond right here in this issue (e.g. by quoting the reviewer feedback as I am doing above ☝️) followed by your response. Does that make sense?

@arfon Thanks! I have copied my replies to here. Hope they are sufficient!
