### 2019-11-13 2c8ac59
modified error message for time matching
