### 2019-11-15 junbinzhao/FluxCalR#2 @jhollist
Nice job!  I just signed off on this over at https://github.com/openjournals/joss-reviews/issues/1751 I will go ahead and close this issue.
