+++ man/FluxCal.Rd
@@ -4,10 +4,23 @@
 \alias{FluxCal}
 \title{Calculate CO2 and CH4 gas fluxes}
 \usage{
-FluxCal(data, win, vol, area, cal = "CO2_CH4", df_cue,
-  cue_type = "End", other = NULL, df_Ta = NULL, ext = 1.5,
-  output = "Flux_output.csv", digits = 3, check_plot = TRUE,
-  ylim_CO2 = NULL, ylim_CH4 = NULL)
+FluxCal(
+  data,
+  win,
+  vol,
+  area,
+  cal = "CO2_CH4",
+  df_cue,
+  cue_type = "End",
+  other = NULL,
+  df_Ta = NULL,
+  ext = 1.5,
+  output = "Flux_output.csv",
+  digits = 3,
+  check_plot = TRUE,
+  ylim_CO2 = NULL,
+  ylim_CH4 = NULL
+)
 }
 \arguments{
 \item{data}{A dataframe generated by the function \code{\link{LoadLGR}} or \code{\link{LoadOther}}.}
