+++ man/SelCue.Rd
@@ -4,8 +4,14 @@
 \alias{SelCue}
 \title{Select the time cues}
 \usage{
-SelCue(data, flux = "CO2", cue = "End", spt = 1, ylim = NULL,
-  save = "Time_cue.csv")
+SelCue(
+  data,
+  flux = "CO2",
+  cue = "End",
+  spt = 1,
+  ylim = NULL,
+  save = "Time_cue.csv"
+)
 }
 \arguments{
 \item{data}{A dataframe generated by the function \code{\link{LoadLGR}} or \code{\link{LoadOther}}.}
