+++ DESCRIPTION
@@ -11,7 +11,7 @@ License: MIT
 Encoding: UTF-8
 LazyData: true
 Imports: dplyr, lubridate, assertthat, magrittr
-RoxygenNote: 6.1.1
+RoxygenNote: 7.0.0
 Suggests: 
     knitr,
     rmarkdown,
