### 2019-11-19 #1751 @ha0ye
@junbinzhao - Great job! Thanks for your hard work in improving the package and responding to our reviews. I have a few minor suggestions to look into:
* increment the version number on the package :)
* as flagged by `goodpractice::gp()` replace `1:nrow(` with `seq_len(nrow(`. This helps avoid errors in the loop if the number of rows is 0.
* do a pass of `devtools::check()` to identify various filenames to cleanup, undocumented arguments, etc. There are some new files that probably should be included in `.Rbuildignore`.
* Upon further inspection, I see that my confusion with the interactive method of selecting timing cues is due to using `x11` on a Mac operating system. (This further gets flagged by `devtools::check()`.) It would be good to dig around examples and documentation for detecting operating system and an interface that is compatible with different user setups.
  - For reference, here is the screenshot from my machine - it doesn't have the stop button, and I had to dig into the documentation for `identify()` to find out how to exit correctly. This was the issue I originally had with trying to interact on my laptop, which doesn't have a secondary mouse button:
> For the usual X11 device the identification process is terminated by pressing any mouse button other than the first. For the quartz device the process is terminated by pressing either the pop-up menu equivalent (usually second mouse button or Ctrl-click) or the ESC key.

![Screen Shot 2019-11-19 at 10 59 15 AM](https://user-images.githubusercontent.com/5835366/69163146-b597ad80-0abb-11ea-94ca-16bc1687799e.png)
