### 2019-11-20 d8feed0
modified code that the timestamp matches the closest point instead of reporting a error
