### 2019-11-20 d6f677c
refined the package according to recommendations of package check
