+++ R/Select_time_cues.R
@@ -4,7 +4,7 @@
 #' clicking the corresponding start or end measurement points on the interactive graphs generated by the function.
 #' The data frame returned from this function can be assigned to the argument \code{df_cue} in the \code{\link{FluxCal}} function.
 #' Note that the time cue data frame can also be prepared by the user independent of this function following the format of
-#' "Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
+#' "_1.csv" and "_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
 #'
 #' @param data A dataframe generated by the function \code{\link{LoadLGR}} or \code{\link{LoadOther}}.
 #'
@@ -81,7 +81,7 @@ SelCue <- function(data,
   # plot flux vs time for locating the peaks and valleys
   In <- c() # create a variable as Index
   for (i in 1:spt){
-    x11(width = 16,height = 10)
+    dev.new(width = 16,height = 10,noRStudioGD=TRUE)
     with(data[c(((i-1)*nr/spt)+1):c(i*nr/spt),],
          plot(Time,flux_plot,
               ylab = ylab,
@@ -110,7 +110,7 @@ SelCue <- function(data,
 
   # output the time cues if necessary
   if (assertthat::is.string(save)){
-    write.csv(Time_q,file = save,row.names = F)
+    utils::write.csv(Time_q,file = save,row.names = F)
   }
   return(Time_q)
 }
