+++ R/Load.R
@@ -6,7 +6,7 @@
 #'
 #' @param file A string contains the directory path and file name of the raw data.
 #' The data should be provided as what is exported from the LGR analyzer.
-#' @param Time_format A string indicates the format of timestamps. Default: "mdy_HMS".
+#' @param time_format A string indicates the format of timestamps. Default: "mdy_HMS".
 #' It also takes the format "dmy_HMS" or "ymd_HMS", depending on the timestamps output from the analyzer.
 #'
 #' @return A data frame with all data in the right format for the functions \code{\link{SelCue}} and \code{\link{FluxCal}}.
@@ -26,7 +26,7 @@ LoadLGR <- function(file,
   # argument check
   time_format <- match.arg(time_format,c("mdy_HMS","dmy_HMS","ymd_HMS"))
   # Load the raw data after deleting the text massages at the beginning and the end of each raw data file generate by LGR
-  flux <- read.table(file, sep= ",", header = TRUE,skip = 1,fill = TRUE,stringsAsFactors = FALSE)
+  flux <- utils::read.table(file, sep= ",", header = TRUE,skip = 1,fill = TRUE,stringsAsFactors = FALSE)
   flux <- subset(flux,!is.na(flux[,2])) # remove the comments generated by LGR
   # replace the time with "lubridated" time with different format
   flux$Time <- try(suppressWarnings(lubridate::parse_date_time(flux$Time,orders = time_format)),
@@ -94,7 +94,7 @@ LoadOther <- function(file,
   # define the pipe from the package "magrittr"
   `%>%` <- magrittr::`%>%`
   # Load the raw data after deleting the text massages at the beginning and the end of each raw data file
-  flux1 <- read.table(file, sep= sep, header = TRUE,skip = skip,fill = TRUE,stringsAsFactors = FALSE,...)
+  flux1 <- utils::read.table(file, sep= sep, header = TRUE,skip = skip,fill = TRUE,stringsAsFactors = FALSE,...)
   flux1 <- subset(flux1,!is.na(flux1[,2])) # remove the comments generated by analyzer
   # replace the time with "lubridated" time with different format
   if (time_format=="HMS"){ # if only time is provided, then combine the time with date first
