+++ man/SelCue.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Select time cues.R
+% Please edit documentation in R/Select_time_cues.R
 \name{SelCue}
 \alias{SelCue}
 \title{Select the time cues}
@@ -45,7 +45,7 @@ This is a function to manually select and extract the start or the end time of e
 clicking the corresponding start or end measurement points on the interactive graphs generated by the function.
 The data frame returned from this function can be assigned to the argument \code{df_cue} in the \code{\link{FluxCal}} function.
 Note that the time cue data frame can also be prepared by the user independent of this function following the format of
-"Time & Ta_1.csv" and "Time & Ta_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
+"_1.csv" and "_2.csv" at https://github.com/junbinzhao/FluxCalR/tree/master/inst/extdata).
 }
 \examples{
 \donttest{
