+++ DESCRIPTION
@@ -10,8 +10,7 @@ Depends: R (>= 3.4.0)
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
-Imports: dplyr, lubridate, assertthat, magrittr, utils
-Remotes: skgrange/threadr
+Imports: dplyr, lubridate, assertthat, magrittr, utils, birk
 RoxygenNote: 7.0.0
 Suggests: 
     knitr,
