### 2019-11-20 7f43ab5
fix the non-binding variable and function notes
