+++ R/globals.R
@@ -1,0 +1,2 @@
+#' @import utils
+utils::globalVariables(c("Date_ms","Index","Slope","Time","Tk","X.CH4.d_ppm","X.CO2.d_ppm"))
