+++ NAMESPACE
@@ -4,3 +4,13 @@ export(FluxCal)
 export(LoadLGR)
 export(LoadOther)
 export(SelCue)
+import(utils)
+importFrom(grDevices,dev.new)
+importFrom(grDevices,dev.off)
+importFrom(graphics,abline)
+importFrom(graphics,axis.POSIXct)
+importFrom(graphics,lines)
+importFrom(graphics,par)
+importFrom(graphics,plot)
+importFrom(graphics,text)
+importFrom(stats,lm)
