### 2019-11-20 5374b92
added instruction for MAC user for cue selections [ci skip]
