+++ paper.md
@@ -38,7 +38,7 @@ There are some programs that calculate the fluxes measured using static chamber 
 
 This R package ``FluxCalR`` is designed to calculate CO~2~ and CH~4~ gas fluxes measured with static chamber method. It provides an easy solution to calculate all the fluxes in one single data file exported by most gas analyzers that includes continuously measured gas concentrations for multiple measurements. Overall, users can complete the flux calculation following three simple steps that implement the build-in functions: 1) load the raw dataset and convert it to the required format; 2) define the time cues that correspond to the start and/or end of the measurements, and 3) calculate the fluxes based on the time cues. 
 
-``FluxCalR`` has several features to facilitate the process of flux calculation. Particularly, it has multiple options to identify valid data for the measurements which greatly improves the flexibility and efficiency of flux calculation. It is also flexible in the inputs (e.g., ambient temperature, gas type) that can meet the needs from measurements with different protocols. It ensures the best quality and representativeness of each flux by choosing the greatest determination coefficient (R^2^) within the defined data range, making the calculated results more reproducible. In addition, the ``FluxCalR`` also provides graphs to visualize the linear regressions on top of the original data to ensure no mismatch occurs. 
+``FluxCalR`` has several features to facilitate the process of flux calculation. Particularly, it has multiple options to identify valid data for the measurements which greatly improves the flexibility and efficiency of flux calculation. It is also flexible in the inputs (e.g., ambient temperature, gas type) that can meet the needs of measurements with different protocols. It ensures the best quality and representativeness of each flux by choosing the regression with the greatest determination coefficient (R^2^) within the defined data range, making the calculated results more reproducible. In addition, the ``FluxCalR`` also provides graphs to visualize the linear regressions on top of the original data to ensure no mismatch occurs. 
 
 ``FluxCalR`` is meant to be used by, for example, ecosystem researchers who carry out greenhouse gas flux measurements using the static chamber method. It simplifies and shortens the process of tedious flux calculations and helps one focus more on data analysis with the ready calculated fluxes. The functions in the package can be integrated into users' data processing/analyzing workflow. At the same time, it also saves the calculated fluxes to a local file that can be loaded by R again later or further processed in other programs. The original functions in the package have been used for flux calculations in the previous publications [@Zhao:2016; @Zhao:2017] and in other ongoing research projects performed in the Everglades.  
 
