### 2019-11-20 #1751 @junbinzhao
@ha0ye Thank you very much for testing the package and giving the comments! Here are how I revised the package:

1. I have changed the version to 0.2.0
2. `1:nrow()` are now all replaced by `seq_len(nrow())`
3. I did the `check` and incorporated all the suggestions. Now it has zero error, warning or notes.
4. I have changed `x()` to `dev.new()` according to the suggestion from the `check` report. Now it should be compatible with systems other than windows as well. Also, I added an additional instruction to the second step in the cue selecting graph "(Mac users press 'Esc')".

Please let me know if you have other concerns or have found other problems.

Thank you!

Junbin
