### 2019-11-21 #1751 @junbinzhao
@arfon Thank you! I have now created a new release and archived the package in Zenodo (DOI: 10.5281/zenodo.3549398). I also went through the paper again and added some funding sources and made a few edits. Please let me know how I can proceed from here.
