### 2019-11-21 780727b
Small fixes.
