+++ paper.md
@@ -24,7 +24,7 @@ bibliography: paper.bib
 
 # Summary
 
-As the main drivers of climate change, greenhouse gas (e.g., CO~2~ and CH~4~) emissions have been monitored intensively across the globe. The static chamber is one of the most commonly used approaches for measuring greenhouse gas fluxes from ecosystems (e.g., stem/soil respiration, CH~4~ emission, etc.) because of its easy implementation, high accuracy and low cost [@Pumpanen:2004]. To perform the measurements, a gas analyzer is usually used to measure the changes of greenhouse gas concentrations within a closed chamber that covers an area of interest (e.g., soil surface) over a certain period of time (usually several minutes). The flux rates (*F*) are, then, calculated from the recorded gas concentrations assuming that the changing rate is linear:
+As the main drivers of climate change, greenhouse gas (e.g., CO~2~ and CH~4~) emissions have been monitored intensively across the globe. The static chamber is one of the most commonly used approaches for measuring greenhouse gas fluxes from ecosystems (e.g., stem/soil respiration, CH~4~ emission, etc.) because of its easy implementation, high accuracy and low cost [@Pumpanen:2004]. To perform the measurements, a gas analyzer is usually used to measure the changes of greenhouse gas concentrations within a closed chamber that covers an area of interest (e.g., soil surface) over a certain period of time (usually several minutes). The flux rates (*F*) are then calculated from the recorded gas concentrations assuming that the changing rate is linear:
 
 $$
 F = vol/(R \cdot Ta \cdot area) \cdot dG/dt
@@ -34,9 +34,9 @@ where *vol* is the volume of the chamber (l), *R* is the universal gas constant 
 
 Even though the method itself appears to be straightforward, to calculate the fluxes from the raw gas concentration data can be tedious, because one usually has to manually extract the data for each measurement from an irregular mix of valid data and unwanted records. After selecting the valid data, it is important to choose a window that delivers the best regression for each measurement to ensure the quality and representativeness of the calculated flux. The R^2^ of the regression is usually taken to determine the quality of a flux measurement and a regression with low R^2^ is usually discarded [@Zhao:2016; @Zhao:2017]. Therefore, a tool that calculates the gas concentration changes into fluxes in an efficient and objective fashion is needed. 
 
-There are some programs that calculate the fluxes measured using static chamber method, such as the Trace Gas Analyzer Platform (LI‐COR Inc., Lincoln, NE, USA), the R package ``flux`` [@Jurasinski:2014], and Flux Puppy [@Carbone:2019]. However, they are either compatible to only one designated system, unable to calculate fluxes for multiple gases, or hard to handle continuously measured gas concentrations. Therefore, there is still a need for a free, flexible and easy-to-use software that is designed to calculate ecosystem gas fluxes from continuously measured gas concentrations. 
+There are some programs that calculate the fluxes measured using static chamber method, such as the Trace Gas Analyzer Platform (LI‐COR Inc., Lincoln, NE, USA), the R package ``flux`` [@Jurasinski:2014], and Flux Puppy [@Carbone:2019]. However, they are either compatible to only one designated system, unable to calculate fluxes for multiple gases, or hard to handle continuously measured gas concentrations. Therefore there is still a need for a free, flexible, and easy-to-use software that is designed to calculate ecosystem gas fluxes from continuously measured gas concentrations. 
 
[…]
