### 2019-11-21 #1751 @arfon
@junbinzhao - I have made a couple of small stylistic changes here: https://github.com/junbinzhao/FluxCalR/pull/4
